% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_manager_and_explore.R
\name{add_db_df}
\alias{add_db_df}
\title{Marker data frame to input data for add.db}
\usage{
add_db_df(mdf = NULL)
}
\description{
Marker data frame to input data for add.db
}
\examples{
mdf <- data.frame(cellName = c('Naive_CD4T','Naive_CD4T','Naive_CD4T','Epithelial','Epithelial'),
geneSymbol = c('CD3D','CD4','CCR7','EPCAM','KRT8'), stringsAsFactors = F)
user_db <- add_db_df(mdf)

}
