% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_manager_and_explore.R
\name{add_db_list}
\alias{add_db_list}
\title{Marker list to input data for add.db}
\usage{
add_db_list(mlist = NULL)
}
\description{
Marker list to input data for add.db
}
\examples{
mlist <- list('Naive CD4 T cell' = c('CD3D','CD4','CCR7'),'Epithelial' = c('EPCAM','KRT8'))
user_db <- add_db_list(mlist)

}
