% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{allde}
\alias{allde}
\title{Find different expression genes.}
\usage{
allde(seurat, group.by = NULL, BPPARAM = SerialParam(), ...)
}
\value{
A dataframe of markers list.
}
\description{
Find different expression genes.
}
\examples{
data('pbmc')
pbmc@meta.data$Group <- Idents(pbmc)
markers <- allde(seurat=pbmc, group.by="Annotation", BPPARAM=MulticoreParam(12))
}
