% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{angrycell_v1.1.7}
\alias{angrycell_v1.1.7}
\title{Find candidate celltypes for each inputed group.}
\usage{
angrycell_v1.1.7(
  object = SeuratS4,
  select.db = c("db_self", "db_lung_zzm"),
  active.assay = "RNA",
  add.db = NULL,
  DE = NULL,
  do.DE = FALSE,
  test.use = "wilcox",
  core.num = 1,
  min.pct = 0.3,
  min.exp = 0,
  show.top = 1,
  use.sigmoid.method = FALSE,
  sigmoid.min = 0.85,
  sigmoid.pct = 0.01,
  HK.model.expr = FALSE,
  HK.model.drop = FALSE,
  HK.model.drop2 = FALSE,
  area.frac = 0.4,
  use.definite.markers = FALSE,
  plot.result = FALSE,
  reduction = "tsne",
  species = "Homosapiens",
  mm10.to.human = F,
  test.method = 2,
  ...
)
}
\arguments{
\item{object}{A seurat object.}

\item{select.db}{Marker genes database for usage. Available options:
db_self(basic database), db_lung_zzm, db_immune_qss, db_cellmarker, panglaodb_hs,
db_ovarian_cancer_T;db_self_mm10(basic database), db_cellmarker_mm10, panglaodb_mm.
Default is basic database + lung_markers_zhangzemin_Cell2019.}

\item{active.assay}{Default assay in seurat object which used to model.}

\item{add.db}{A dataframe. Add database by user. Colnames is c("cellName", "geneSymbol").
each geneset in "geneSymbol" is a character, such as 'CD3D,CD4,CCR7'.}

\item{DE}{The result of FindAllMarker as candidate genes.}

\item{do.DE}{Find defferent expression genes as candidate genes.}

\item{test.use}{Denotes which test to use if do.DE is TRUE.
Available options are:
\itemize{
 \item{"wilcox"} : Identifies differentially expressed genes between two
 groups of cells using a Wilcoxon Rank Sum test (default)
 \item{"t"} : Identify differentially expressed genes between two groups of
 cells using the Student's t-test.
}}

\item{core.num}{Number of cores used for do.DE. Default is 1}

\item{min.pct}{Only candidate genes that are detected in a minimum fraction of
min.pct cells. Default is 0.3}

\item{min.exp}{Only detected genes with expression level higher than min.exp.}

\item{show.top}{To show top n candidate predicted celltypes from the result of
GO enrichment. Default is 1}

\item{use.sigmoid.method}{Another available method for cadidate genes selection.
Using sigmoid function to identify candidate genes instead of min.pct.}

\item{sigmoid.min}{If use.sigmoid.method is TURE, only cells that are expressing
the detected gene in a minimum fraction of sigmoid.min sigmoid value.}

\item{sigmoid.pct}{If use.sigmoid.method is TURE, only genes that are candidate
genes in a minimum fraction of sigmoid.pct detected cells in one cluster.}

\item{HK.model.expr}{Another available method for cadidate genes selection. Get the
cutoff value base on the expression of house keep genes. The cutoff value from the
intersection point in a minimum fraction of area.frac area under curve
(dropOutRate x Expression).}

\item{HK.model.drop}{Another available method for cadidate genes selection. Get the
cutoff value base on the DropOutRate of house keep genes. The cutoff value from the
intersection point in a minimum fraction of area.frac area under curve
(dropOutRate x Expression).}

\item{HK.model.drop2}{Another available method for cadidate genes selection.
The median value of the DropOutRate of house keep genes as cutoff value.}

\item{area.frac}{Minimum fraction of area under a curve from HK.model.}

\item{use.definite.markers}{Use strong marker to identify cell types. Default is F.}

\item{plot.result}{Plot predicted cell type and marker genes in GO terms.}

\item{reduction}{Select the reduction method.}

\item{species}{Optional speices: Homosapiens, Mouse.}

\item{mm10.to.human}{Transfer mouse homolog genes to human.}

\item{test.method}{Methods for hepergeometric test.
\itemize{
 \item{1}: lapply;
 \item{2}: phyper;
 \item{3}: Rcpp Ramanujan's approximation.
 }}

\item{...}{Arguments passed to signDE.}
}
\value{
A dataframe of celltype annotation.
}
\description{
Find candidate celltypes for each inputed group.
}
\examples{
angrycell(pbmc) #pbmc is a seurat object.
angrycell(pbmc, min.pct = 0.3, show.top = 10, use.definite.markers = F,
reduction = 'umap')
angrycell(pbmc, min.pct = 0.3, show.top = 10, use.definite.markers = TRUE,
plot.result = TRUE, reduction = 'umap')
angrycell(pbmc, min.pct = 0.3, show.top = 10, use.sigmoid.method = TRUE,
sigmoid.min = 0.8, sigmoid.pct = 0.01)
angrycell(pbmc, HK.model.expr = TRUE, area.frac = 0.4)
angrycell(pbmc, HK.model.drop = TRUE, area.frac = 0.1)
angrycell(pbmc, HK.model.drop2 = TRUE)
}
