% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_manager_and_explore.R
\name{angrycelldb}
\alias{angrycelldb}
\title{Search markers or cell types.}
\usage{
angrycelldb(genename = NULL, cellname = NULL, use.specific.marker = F)
}
\description{
Search markers or cell types.
}
