% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{anno_AUCell}
\alias{anno_AUCell}
\title{Annotate Celltypes by AUCell}
\usage{
anno_AUCell(
  object,
  species,
  assay = "RNA",
  raw_cluster = NULL,
  ensembl_version = 98,
  label_colname = "AUCell_label_cell",
  label_raw_cluster_colname = "AUCell_label_raw_cluster",
  scsig = NULL,
  scsig.subset = c("_", "Cord_Blood", "Esophagus", "Stomach", "Small_Intestine",
    "Large_Intestine", "PFC", "Embryonic", "Midbrain", "Bone_Marrow", "Liver",
    "Fetal_Kidney", "Adult_Kidney", "Pancreas"),
  nCores = 4,
  ...
)
}
\description{
Annotate Celltypes by AUCell
}
\examples{
pbmc <- anno_AUCell(pbmc, species = 'Human', scsig.subset = 'Bone_Marrow') # based on individual cells
pbmc <- anno_AUCell(pbmc, species = 'Human', , scsig.subset = 'Bone_Marrow', raw_cluster = 'seurat_clusters') # based on seurat clusters
}
