% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{anno_ellmer}
\alias{anno_ellmer}
\title{Annotate cell types using an AI wrapper embedded in the ellmer R package.}
\usage{
anno_ellmer(
  deg = NULL,
  genelist = NULL,
  tissuename = NULL,
  base_url = "http://chatapi.littlewheat.com/v1",
  api_key = "sk-HgtySiUAhSLiZTlDRhNE7aEbERJOuSumUveDxYfAUy8YvDfM",
  model = "gpt-3.5-turbo",
  llm_function = c("openai", "deepseek", "ollama"),
  ollama_model = "llama3.2",
  prompts = NULL,
  prompts.add = NULL,
  rm_str = c("\\\\*"),
  sep = c("\\\\: ", "- ", "\\\\* ", "\\\\. "),
  as.order = F,
  return.prompt = F,
  return.answer = F,
  seed = 1234
)
}
\description{
Annotate cell types using an AI wrapper embedded in the ellmer R package.
}
\examples{
Idents(pbmc) <- pbmc$seurat_clusters
markers <- FindAllMarkers(pbmc, logfc.threshold = 0.5, test.use = 'MAST', only.pos = T)
top10 <- markers \%>\% group_by(cluster) \%>\% top_n(10, avg_log2FC)
annodf <- anno_ellmer(top10)
# depends on local model ollama. First download and install Ollama, then install model with ollama pull llama3.2. Test on Mac.
annodf <- anno_ellmer(top10, llm_function = 'ollama')
}
