% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{anno_llm}
\alias{anno_llm}
\title{Annotate cell types using LLM model and DEGs}
\usage{
anno_llm(
  object,
  deg,
  raw_cluster,
  label_raw_cluster_colname,
  llm_function,
  base_url,
  api_key,
  ...
)
}
\description{
Annotate cell types using LLM model and DEGs
}
