% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{anno_openai}
\alias{anno_openai}
\title{Annotate celltypes by ChatGPT.}
\usage{
anno_openai(
  deg = NULL,
  genelist = NULL,
  tissuename = NULL,
  base_url = "http://chatapi.littlewheat.com/v1",
  api_key = "sk-HgtySiUAhSLiZTlDRhNE7aEbERJOuSumUveDxYfAUy8YvDfM",
  model = "gpt-3.5-turbo",
  seed = 123
)
}
\description{
Annotate celltypes by ChatGPT.
}
\examples{
Idents(pbmc) <- pbmc$seurat_clusters
markers <- FindAllMarkers(pbmc, logfc.threshold = 0.5, test.use = 'MAST', only.pos = T)
top10 <- markers \%>\% group_by(cluster) \%>\% top_n(10, avg_log2FC)
annodf <- anno_openai(top10)
}
