% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{anno_SingleR}
\alias{anno_SingleR}
\title{Annotate Celltypes by SingleR}
\usage{
anno_SingleR(
  object,
  species,
  assay = "RNA",
  raw_cluster = NULL,
  ensembl_version = 98,
  label_colname = "SingleR_label_cell",
  label_raw_cluster_colname = "SingleR_label_raw_cluster",
  ref = NULL,
  ref.label = c("label.main", "label.fine", "label.ont"),
  ...
)
}
\description{
Annotate Celltypes by SingleR
}
