% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_top_gene.R
\name{anno_top_gene_CODEX}
\alias{anno_top_gene_CODEX}
\title{Annotate raw clusters by max expression, and return an object with annotation Idents}
\usage{
anno_top_gene_CODEX(
  object,
  markerlist,
  markerdf = NULL,
  group.by = "seurat_clusters",
  cut.off = 0.95,
  assign.noise = F
)
}
\description{
Annotate raw clusters by max expression, and return an object with annotation Idents
}
\examples{
markerdf <- data.frame(celltypes = c('T','T','NK','NK','Mono', 'Mono','DC','DC','B','B','Platelet'), 
markers = c('CD3D','CD3E','NCAM1','NKG7','CD14','FCGR3A','CST3','CD1C','CD79A','MS4A1','PPBP'))
colnames(markerdf)
#[1] "celltypes" "markers"
pbmc_anno <- anno_top_gene(pbmc, markerdf, group.by = 'seurat_clusters')
DimPlot(pbmc_anno, label = T)

}
