% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annocell}
\alias{annocell}
\title{Annotate Celltypes in one function}
\usage{
annocell(
  object,
  species,
  assay = "RNA",
  raw_cluster = NULL,
  method = c("SingleR"),
  ensembl_version = 98,
  label_colname = "SingleR_label_cell",
  label_raw_cluster_colname = "SingleR_label_raw_cluster",
  ref = NULL,
  ref.label = c("label.main", "label.fine", "label.ont"),
  ...
)
}
\description{
Annotate Celltypes in one function
}
