% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchR_plot_highlight.R
\name{ArchR_highallcl}
\alias{ArchR_highallcl}
\title{highlight clusters}
\usage{
ArchR_highallcl(
  object,
  colorBy = "cellColData",
  name = "Clusters",
  size = 0.1,
  size.title = 10,
  embedding = "UMAP",
  ncol = NULL
)
}
\description{
highlight clusters
}
