% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchR_plot_pseudo.R
\name{ArchR_merge_pseudo}
\alias{ArchR_merge_pseudo}
\title{Merge pseudo paths on one plot}
\usage{
ArchR_merge_pseudo(
  object,
  traj_namelist,
  point.name,
  color.line = NULL,
  color.point = NULL
)
}
\description{
Merge pseudo paths on one plot
}
