% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchR_plot_highlight.R
\name{ArchR_one_cl}
\alias{ArchR_one_cl}
\title{highlight cluster}
\usage{
ArchR_one_cl(
  object,
  cl,
  colorBy = "cellColData",
  name = "Clusters",
  size = 0.1,
  size.title = 10,
  embedding = "UMAP"
)
}
\description{
highlight cluster
}
