% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{aveheatmap}
\alias{aveheatmap}
\title{heatmap of average expression}
\usage{
aveheatmap(
  object,
  paths,
  ident = Idents(object),
  assay = "RNA",
  anno_col = NULL,
  anno_row = NULL
)
}
\description{
heatmap of average expression
}
