% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{bigdataTOmat}
\alias{bigdataTOmat}
\title{In order to avoid memory problem, Convert large dgCMatrix to Matrix by separation.}
\usage{
bigdataTOmat(
  bigdata,
  min_num = 10000,
  transpose = F,
  as.type = c("matrix", "TsparseMatrix", "dgCMatrix")
)
}
\description{
In order to avoid memory problem, Convert large dgCMatrix to Matrix by separation.
}
