% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_score.R
\name{box_score}
\alias{box_score}
\title{Box plot of gene score.}
\usage{
box_score(object, geneset, genesetname, cellpos, AddModuleScore = F)
}
\description{
Box plot of gene score.
}
\examples{
cellpos <- pbmc$Annotation \%in\% c('Naive CD4 T', 'Memory CD4 T','CD8 T','NK')
geneset <- c('CCR7','TCF7','SELL')
box_score(pbmc, geneset,'Naive lymphocytes',cellpos)
geneset <- c('GATA3','PPARG','PTGDR2','IL4','IL5','IL13','CSF1','IL1RL1','IL4R',
'IL17RB','CCR4','CYSLTR1','HPGDS','TESPA1','AHR')
box_score(pbmc, geneset,'Type 2 inflammation in Th2 and ILC2',cellpos)

}
