% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_box.R
\name{boxplot}
\alias{boxplot}
\title{boxplot}
\usage{
boxplot(
  data = NULL,
  paths = NULL,
  ident = NULL,
  meltdata = NULL,
  facet.toward = "row",
  do.log = F,
  strip.text.size = 8,
  color.use = NULL,
  jitter = F,
  cell.order = NULL,
  add.ave.point = F,
  add.line = F,
  line.size = 0.5,
  add.point = T,
  mid.color.point = "auto",
  legend.position = "top",
  legend.name = "",
  jitter.size = 0.1,
  text.x.size = 8,
  text.y.size = 8,
  hjust = 0,
  vjust = 0,
  legend.title.size = 8,
  axis.text.x.angle = NULL,
  do.test = F,
  comparisons = NULL,
  test.method = c("wilcox.test", "t.test", "anova", "kruskal.test")[1]
)
}
\description{
boxplot
}
