% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_score.R
\name{boxplot_score}
\alias{boxplot_score}
\title{Box plot of gene score.}
\usage{
boxplot_score(
  object,
  assay = "RNA",
  features,
  group.by = NULL,
  select.celltypes = NULL,
  color.use = NULL,
  test.use = c("none", "t", "wilicox"),
  adjust.p = T,
  plabel = c("p.adj.signif", "p.adj", "p"),
  adjust.method = c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr",
    "none"),
  return.stat = F
)
}
\description{
Box plot of gene score.
}
\examples{
cellt <- c('Naive CD4 T', 'Memory CD4 T','CD8 T','NK')
geneset <- c('CCR7','TCF7','SELL')
boxplot_score(pbmc, features = geneset,select.celltypes = cellt)
# Show number of star of p value.
boxplot_score(pbmc, features = geneset,select.celltypes = cellt, 
test.use = 'wilicox', adjust.method = 'bonferroni')
# Show p value
boxplot_score(pbmc, features = geneset,select.celltypes = cellt, 
test.use = 'wilicox', adjust.method = 'bonferroni', plabel = 'p.adj')
# Return statistic data frame
df <- boxplot_score(pbmc, features = geneset,select.celltypes = cellt, 
test.use = 'wilicox', adjust.method = 'bonferroni', return.stat = T)

}
