% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_score.R
\name{cal_cnvScore}
\alias{cal_cnvScore}
\title{Calculate CNV score}
\usage{
cal_cnvScore(
  infercnv_obj = NULL,
  genefile.path = NULL,
  do.kmeans = FALSE,
  k = 7,
  do.correlation = FALSE,
  percent_top = 5,
  out.dir = ".",
  out.prefix = "cnvScore",
  save.heatmap = F,
  min.cnv = 0.4,
  max.cnv = 1.2,
  plot = F,
  return.cordata = T
)
}
\description{
Calculate CNV score
}
