% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_process.R
\name{cbind_fill}
\alias{cbind_fill}
\title{Combine multi vectors in different lengths, filled with NA.}
\usage{
cbind_fill(rawlist)
}
\arguments{
\item{rawlist}{A list contains multi vectors in which elements have names.}
}
\value{
A dataframe.
}
\description{
Combine multi vectors in different lengths, filled with NA.
}
\examples{
a <- c(1,3,5)
b <- c(1,2,3,4)
c <- c(1,2,4,5)
names(a) <- c('a','c','e')
names(b) <- c('a','b','c','d')
names(c) <- c('a','b','c','e')
rawlist <- list(a,b,c)
data <- cbind_fill(rawlist)
}
