% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion_statistics.R
\name{cell_t_test}
\alias{cell_t_test}
\title{T-test of cell number}
\usage{
cell_t_test(
  metadf,
  Group = "Group",
  Annotation = "Annotation",
  Sample = "Sample",
  alternative = c("two.sided", "great", "less")
)
}
\description{
T-test of cell number
}
