% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion_statistics.R
\name{cell_test}
\alias{cell_test}
\title{Statistics of cell proportion}
\usage{
cell_test(
  dataset,
  Group = "Group",
  Annotation = "Annotation",
  Sample = "Sample",
  test = c("fisher", "chisq", "t", "wilcox"),
  model = c("2x2", "pairs"),
  alternative = c("two.sided", "great", "less"),
  adjust.method = c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr",
    "none")
)
}
\description{
Statistics of cell proportion
}
\examples{
# 2x2
statdf <- cell_test(meta200, Group = 'Group',Annotation = 'Annotation',
Sample = 'Sample', model = '2x2', test = 'fisher',alternative = 'two.sided',
adjust.method = 'BH')
# pairwise
statdf <- cell_test(meta200, Group = 'Group',Annotation = 'Annotation',
Sample = 'Sample', model = '2x2', test = 'fisher',alternative = 'two.sided',
adjust.method = 'BH')

}
