% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_enrichment.R
\name{changeID}
\alias{changeID}
\title{Function for gene ID transform}
\usage{
changeID(
  genelist,
  species = c("Homosapiens", "mm10")[1],
  biomart = "ENSEMBL_MART_ENSEMBL",
  host = "dec2016.archive.ensembl.org",
  keyType = c("SYMBOL", "ENTREZID", "ENSEMBL")[3],
  toType = c("SYMBOL", "ENTREZID", "ENSEMBL")[1]
)
}
\description{
Function for gene ID transform
}
