% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{check_replace_name}
\alias{check_replace_name}
\title{Check whether there is special symbol in names of cell barcode or annotation to avoid error.}
\usage{
check_replace_name(celltype, pattern = NULL, replacement = NULL, fixed = F)
}
\arguments{
\item{celltype}{A vector, can be barcode or annotation.}
}
\description{
Check whether there is special symbol in names of cell barcode or annotation to avoid error.
}
