% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_enrichKEGG_GSEA_GO.R,
%   R/pathway_enrichment.R
\name{clusterGO}
\alias{clusterGO}
\title{GO over-representation analysis}
\usage{
clusterGO(
  markerlist,
  use.top.marker = 100,
  filter.gene.pattern = NULL,
  ont = "BP",
  pAdjustMethod = "BH",
  pvalueCutoff = 0.05,
  qvalueCutoff = 0.2,
  keyType = "SYMBOL",
  OrgDb = "org.Hs.eg.db",
  n.cores = 1,
  ...
)

clusterGO(
  markerlist,
  use.top.marker = 100,
  filter.gene.pattern = NULL,
  ont = "BP",
  pAdjustMethod = "BH",
  pvalueCutoff = 0.05,
  qvalueCutoff = 0.2,
  keyType = "SYMBOL",
  OrgDb = "org.Hs.eg.db",
  n.cores = 1,
  ...
)
}
\arguments{
\item{markerlist}{A DEG dataframe from Seurat::FindAllMarkers.}

\item{use.top.marker}{Number of top markers for GO enrichment. Default is 100.}

\item{filter.gene.pattern}{Filtering gene.}

\item{ont}{Ontology. Parameter in clusterProfiler::enrichGO.}

\item{pAdjustMethod}{Parameter in clusterProfiler::enrichGO.}

\item{pvalueCutoff}{Parameter in clusterProfiler::enrichGO.}

\item{qvalueCutoff}{Parameter in clusterProfiler::enrichGO.}

\item{keyType}{Parameter in clusterProfiler::bitr.}

\item{...}{Parameters in enrichGO.}
}
\description{
GO over-representation analysis

GO enrichment.
GO enrichment by clusterProfiler::enrichGO.
}
\examples{
markerlist <- FindAllMarkers(pbmc,only.pos = TRUE,logfc.threshold = 0.5)
GOmat <- clusterGO(markerlist)
}
