% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_enrichKEGG_GSEA_GO.R
\name{clusterGseKEGG}
\alias{clusterGseKEGG}
\title{Gene set enrichment analysis}
\usage{
clusterGseKEGG(
  markerlist,
  gene_colname = "gene",
  log2FC_colname = "avg_log2FC",
  filter.gene.pattern = NULL,
  geneType = "SYMBOL",
  OrgDb = "org.Hs.eg.db",
  organism = "hsa",
  minGSSize = 10,
  maxGSSize = 500,
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  verbose = TRUE,
  seed = FALSE,
  return.gsea_obj = FALSE,
  by = "fgsea",
  ...
)
}
\description{
Gene set enrichment analysis
}
