% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_enrichKEGG_GSEA_GO.R,
%   R/pathway_enrichment.R
\name{clusterKEGG}
\alias{clusterKEGG}
\title{KEGG pathway over-representation analysis}
\usage{
clusterKEGG(
  markerlist,
  use.top.marker = 100,
  organism = "hsa",
  keyType = "kegg",
  filter.gene.pattern = NULL,
  pAdjustMethod = "BH",
  pvalueCutoff = 0.05,
  qvalueCutoff = 0.2,
  ...
)

clusterKEGG(
  markerlist,
  use.top.marker = 100,
  organism = "hsa",
  keyType = "kegg",
  filter.gene.pattern = NULL,
  pAdjustMethod = "BH",
  pvalueCutoff = 0.05,
  qvalueCutoff = 0.2,
  ...
)
}
\arguments{
\item{markerlist}{A DEG dataframe from Seurat::FindAllMarkers.}

\item{use.top.marker}{Number of top markers for KEGG enrichment. Default is 100.}

\item{filter.gene.pattern}{Filtering gene.}

\item{pAdjustMethod}{Parameter in clusterProfiler::enrichKEGG.}

\item{pvalueCutoff}{Parameter in clusterProfiler::enrichKEGG.}

\item{qvalueCutoff}{Parameter in clusterProfiler::enrichKEGG.}

\item{...}{Parameters in enrichKEGG.}
}
\description{
KEGG pathway over-representation analysis

KEGG enrichment. KEGG enrichment by clusterProfiler::enrichKEGG.
}
\examples{
markerlist <- FindAllMarkers(pbmc,only.pos = TRUE,logfc.threshold = 0.5)
KEGGmat <- clusterKEGG(markerlist)

}
