% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{color_generator}
\alias{color_generator}
\title{Generate colors vector}
\usage{
color_generator(n, palette = 1, color.ramps = F)
}
\arguments{
\item{n}{number of colors.}

\item{palette}{palette of colors. 3 palettes are provided in function: 1,2,3. Or input a vector by user.}

\item{color.ramps}{To extend color if n is larger than palette.}
}
\description{
Generate colors vector
}
