% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_cover_outlier.R
\name{cover_outlier}
\alias{cover_outlier}
\title{Adjust layer of point to cover outliers.}
\usage{
cover_outlier(
  object,
  annotation,
  seurat_clusters = "seurat_clusters",
  reduction = "umap"
)
}
\value{
Ordered point names.
}
\description{
Adjust layer of point to cover outliers.
}
