% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_process.R
\name{data_stat}
\alias{data_stat}
\title{Statistc of cell number before and after cell filtering.}
\usage{
data_stat(orig_metadata, input_gmin, do.plot = FALSE)
}
\arguments{
\item{orig_metadata}{A metadata must contain information: nFeature_RNA,
percent_mt, nCount_RNA and Group.}

\item{input_gmin}{Munimum gene number for filtering cell.}
}
\value{
A dataframe of cell number in different filtering condition; hist plot of cutoff.
}
\description{
Statistc of cell number before and after cell filtering. mt_cut_off: 25, 15, 10, 5%
}
\examples{
samples <- c(rep(c('sample1','sample2','sample3'),(ncol(pbmc) -1)/3),'sample1')
pbmc@meta.data$Sample <- samples
base_stat <- data_stat(pbmc@meta.data, 100)
}
