% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{DEG_All_limma}
\alias{DEG_All_limma}
\title{Running DEG of all clusters using limma. For each cluster, 
cells are compared with cells from all other clusters.}
\usage{
DEG_All_limma(
  object,
  group.colname,
  only.pos = T,
  logFC.threhold = 0.25,
  adj.P.Val = 1,
  batch.colname = NULL,
  n.cores = 4,
  ...
)
}
\arguments{
\item{...}{Parameter to GetAssayData.}
}
\description{
Running DEG of all clusters using limma. For each cluster, 
cells are compared with cells from all other clusters.
}
\examples{
# includes random effects
pbmc$batch <- c(rep('sample1',1000), rep('sample2',1000),rep('sample3',638))
DEG <- DEG_All_limma(pbmc, 'Annotation', batch.colname = 'batch', n.cores = 8)
}
