% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_DAseq.R
\name{DEG_DAseq}
\alias{DEG_DAseq}
\title{Pairwise DEG using DAseq}
\usage{
DEG_DAseq(
  object,
  groupname,
  ident.1,
  ident.2,
  lambda = 1.5,
  n.cores = 4,
  return.model = T,
  da.regions.to.run = NULL,
  p.adjust.method = "fdr",
  min.pct = 0.2,
  logfc.threshold = 0,
  min.cell = 50,
  python2use = "/media/london_B/huangzh/software/miniconda3/envs/FR-Perturb/bin/python",
  GPU = 4,
  reduction = "pca",
  ...
)
}
\arguments{
\item{groupname}{Colname including two groups for DEG analysis, e.g. SNP and WT labels.}

\item{lambda}{numeric, regularization parameter that weights the number of selected genes, a larger lambda leads to fewer genes, default 1.5}

\item{...}{Parameter to GetAssayData.}

\item{batchname}{Colname including covariate in meta.data.}

\item{k.vector}{k values to create the score vector.}

\item{k.folds}{integer, number of data splits used in the neural network, default 10.}

\item{n.runs}{integer, number of times to run the neural network to get the predictions, default 5.}

\item{n.rand}{integer, number of random permutations to run, default 2.}

\item{pred.thres.range}{length-2 vector, top and bottom threshold on DA measure, default NULL, select significant DA cells based on permutation}

\item{deg.func}{STG or Seurat. STG: Use STG (stochastic gates) to select genes that separate each DA region from the rest of the cells.}
}
\description{
Pairwise DEG using DAseq
}
\examples{
packageVersion('Seurat')
# ‘4.4.0’
packageVersion('Matrix')
# '1.6.2'
pbmc <- readRDS('../data/pbmc_v4.rds')
pbmc <- subset(pbmc, Annotation \%in\% c('B','NK'))
pbmc$Annotation <- factor(pbmc$Annotation, levels = unique(pbmc$Annotation))
# includes random effects
pbmc$batch <- c(rep('sample1',100), rep('sample2',100),rep('sample3',299))
DEG <- DEG_DAseq(pbmc, groupname = 'Annotation', ident.1 = 'B', ident.2 = 'NK')

}
