% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_DESeq2.R
\name{DEG_DESeq2}
\alias{DEG_DESeq2}
\title{Pairwise DEG using DEseq2 adjusted for single-cell data.
Refer to https://bioconductor.org/packages/devel/bioc/vignettes/DESeq2/inst/doc/DESeq2.html#recommendations-for-single-cell-analysis. https://hbctraining.github.io/DGE_workshop_salmon_online/lessons/05b_wald_test_results.html}
\usage{
DEG_DESeq2(
  object,
  groupname,
  ident.1,
  ident.2,
  batchname = NULL,
  assay = "RNA",
  min.pct = 0.2,
  logfc.threshold = 0,
  filtergene.rowSums = 0,
  test.method = c("LRT", "Wald"),
  fitType = c("glmGamPoi"),
  useT = F,
  minReplicatesForReplace = Inf,
  calucateSumFactors = F,
  minmu = if (fitType == "glmGamPoi") 1e-06 else 0.5,
  adjust.method = c("bonferroni", "holm", "hochberg", "hommel", "BH", "BY", "fdr",
    "none"),
  parallel = F,
  n.cores = 4,
  alpha = 0.9999,
  cooksCutoff = Inf,
  independentFiltering = F,
  ...
)
}
\arguments{
\item{fitType}{Parameter in function DESeq(). Either "parametric", "local", "mean", or "glmGamPoi" for the type of fitting of dispersions to the mean intensity. See estimateDispersions for description.}

\item{useT}{Parameter in function DESeq(). Logical, passed to nbinomWaldTest, default is FALSE, where Wald statistics are assumed to follow a standard Normal.}

\item{minReplicatesForReplace}{Parameter in function DESeq(). The minimum number of replicates required in order to use replaceOutliers on a sample. If there are samples with so many replicates, the model will be refit after these replacing outliers, flagged by Cook's distance. Set to Inf in order to never replace outliers. It set to Inf for fitType="glmGamPoi".}

\item{minmu}{Lower bound on the estimated count for fitting gene-wise dispersion and for usage with nbinomWaldTest and nbinomLRT. If fitType="glmGamPoi", then 1e-6 will be used (as this fitType is optimized for single cell data, where a lower minmu is recommended), otherwise the default value as evaluated on bulk datasets is 0.5.}

\item{...}{Parameter to GetAssayData.}

\item{test}{Likelihood ratio test (LRT) or Wald test.}
}
\description{
Pairwise DEG using DEseq2 adjusted for single-cell data.
Refer to https://bioconductor.org/packages/devel/bioc/vignettes/DESeq2/inst/doc/DESeq2.html#recommendations-for-single-cell-analysis. https://hbctraining.github.io/DGE_workshop_salmon_online/lessons/05b_wald_test_results.html
}
\examples{
pbmc <- readRDS('../data/pbmc_v4.rds')
pbmc <- subset(pbmc, Annotation \%in\% c('B','NK'))
pbmc$Annotation <- factor(pbmc$Annotation, levels = unique(pbmc$Annotation))
# includes random effects
pbmc$batch <- c(rep('sample1',100), rep('sample2',100),rep('sample3',299))
DEG_LRT <- DEG_DESeq2(pbmc, groupname = 'Annotation', useT = T, test.method = 'LRT', batchname  = 'batch')
DEG_Wald <- DEG_DESeq2(pbmc, groupname = 'Annotation', test.method = 'Wald', parallel = T, fitType = 'parametric', batchname  = 'batch')

}
