% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_edgeR.R
\name{DEG_edgeR}
\alias{DEG_edgeR}
\title{Pairwise DEG using edegeR. 
QLF: Conduct genewise quasi F-tests for a given coefficient or coefficient contrast.
LRT: conduct likelihood ratio tests for a given coefficient or coefficient contrast.}
\usage{
DEG_edgeR(
  object,
  groupname,
  ident.1,
  ident.2,
  batchname = NULL,
  min.pct = 0.2,
  logfc.threshold = 0,
  test = c("LRT", "QLF"),
  adjust.method = c("bonferroni", "holm", "hochberg", "hommel", "BH", "BY", "fdr",
    "none"),
  ...
)
}
\arguments{
\item{test}{likelihood ratio test (LRT) or quasi-likelihood F-test (QLF)}

\item{...}{Parameter to GetAssayData.}

\item{min.cells}{the minimum number of expressed cells for each genes.}
}
\description{
Pairwise DEG using edegeR. 
QLF: Conduct genewise quasi F-tests for a given coefficient or coefficient contrast.
LRT: conduct likelihood ratio tests for a given coefficient or coefficient contrast.
}
\examples{
pbmc <- readRDS('../data/pbmc_v4.rds')
pbmc <- subset(pbmc, Annotation \%in\% c('B','NK'))
pbmc$Annotation <- factor(pbmc$Annotation, levels = unique(pbmc$Annotation))
# includes other covariates
pbmc$batch <- c(rep('sample1',100), rep('sample2',100),rep('sample3',299))
DEG <- DEG_edgeR(pbmc, groupname = 'Annotation', ident.1 = 'B', ident.2 = 'NK', batchname  = c('batch','percent.mt'))
DEG <- DEG_edgeR(pbmc, groupname = 'Annotation', ident.1 = 'B', ident.2 = 'NK', batchname  = c('batch','percent.mt'), test = 'QLF')
}
