% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_FRperturb.R
\name{DEG_FRperturb}
\alias{DEG_FRperturb}
\title{DEG using FR-perturb}
\usage{
DEG_FRperturb(
  object,
  assay_crispr = "crispr_thred",
  workdir = "./",
  run_FR_Perturb_script = "/media/london_B/huangzh/software/FR-Perturb/run_FR_Perturb.py",
  python2use = "/media/london_B/huangzh/software/miniconda3/envs/FR-Perturb/bin/python",
  num.perms = NULL,
  covariates = NULL,
  control.prefix = NULL
)
}
\arguments{
\item{num.perms}{specifies the number of permutations to perform for permutation testing. Default 1000.}
}
\description{
DEG using FR-perturb
}
\examples{
hmoi <- readRDS('../data/obj_pct01_logfc01.rds')
DEG_FRperturb(hmoi, assay_crispr = 'crispr_thred', workdir = "/media/london_B/huangzh/project/benchmark_perturb/TEMP/", control.prefix = 'NTC', covariates = c('percent.mt', 'Sample', 'number_positive_sgRNA'))

}
