% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{DEG_glmmTMB}
\alias{DEG_glmmTMB}
\title{Pairwise DEG using glmmTMB to account for random effect.}
\usage{
DEG_glmmTMB(
  object,
  group.colname,
  ident.1,
  ident.2,
  batch.colname = NULL,
  min.ave.counts = 3,
  adjust.p.method = "fdr"
)
}
\description{
Refer to https://github.com/kdzimm/PseudoreplicationPaper/blob/master/Type_1_Error/Type%201%20-%20Tweedie%20GLMM.Rmd
}
\examples{
pbmc$batch <- c(rep('sample1',1000), rep('sample2',1000),rep('sample3',638))
subs4 <- subset(pbmc, Annotation \%in\% c('CD14+ Mono','FCGR3A+ Mono'))
MASTdf <- DEG_MAST_RE(subs4, 'Annotation', ident.1 = 'CD14+ Mono', ident.2 = 'FCGR3A+ Mono', batch.colname = 'batch',freq_expressed = 0.2)
MASTdf <- MASTdf[MASTdf$FDR < 0.05 & MASTdf$avg_logFC > 0.15,]
}
