% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{DEG_group_per_celltype}
\alias{DEG_group_per_celltype}
\title{DEG of pairwise groups per cell type.}
\usage{
DEG_group_per_celltype(
  object,
  celltype,
  group,
  test.use = "wilcox",
  min.pct = 0.25,
  logfc.threshold = 0.2,
  adj.P.Val = 1,
  batch.colname = NULL,
  mc.cores = 8,
  ...
)
}
\description{
DEG of pairwise groups per cell type.
}
\examples{
samples <- c(rep(c('sample1','sample2','sample3'),(ncol(pbmc) -1)/3),'sample1')
pbmc@meta.data$Sample <- samples
DEGlist <- DEG_group_per_celltype(pbmc, celltype = 'Annotation',group = 'Sample',mc.cores = 1)
openxlsx::write.xlsx(DEGlist, file = './DEG_groups_per_celltype.xlsx')

}
