% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_GSFA.R
\name{DEG_GSFA}
\alias{DEG_GSFA}
\title{DEG using GSFA}
\usage{
DEG_GSFA(
  object,
  G_mat,
  covariates,
  num_top_genes = 6000,
  K = 20,
  prior_type = "mixture_normal",
  init.method = "svd",
  niter = 3000,
  used_niter = 1000,
  seed = 14314,
  ...
)
}
\arguments{
\item{G_mat}{A binary matrix assigning sgRNA for each cell. Each cell contains a unique gRNA. 1 or 0. Column is cell.}

\item{K}{Number of factors to use in the model}

\item{prior_type}{Type of sparse prior used on gene weights, 
can be "mixture_normal" or "spike_slab", "mixture_normal" sometimes works better in inducing sparsity.}

\item{init.method}{Method to initialize the factors, can be one of "svd" (truncated SVD on ‘Y’) or "random"; description}

\item{...}{Parameter to GetAssayData.}
}
\description{
DEG using GSFA
}
\examples{
## source conda environment GSFA
# source activate GSFA 
obj <- readRDS('../data/obj_pct01_logfc01.rds')
G_mat <- t(as.matrix(obj@assays$crispr_thred@counts))
obj$Sample <- factor(obj$Sample, levels = unique(obj$Sample))
fitdata <- DEG_GSFA(obj, gmat, batchname = 'Sample')

}
