% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{DEG_limma_trend}
\alias{DEG_limma_trend}
\title{Pairwise DEG using limma-trend
# citation: https://ucdavis-bioinformatics-training.github.io/2019-single-cell-RNA-sequencing-Workshop-UCD_UCSF/scrnaseq_analysis/scRNA_Workshop-PART6.html}
\usage{
DEG_limma_trend(
  object,
  group.colname,
  ident.1,
  ident.2,
  batch.colname = NULL,
  min.cell = 1,
  calcNormFactors = F,
  max.cpm.cut.off = 1,
  prior.count = 2,
  duplicateCorrelation.name = NULL,
  trend = F,
  sort.by = "P",
  ...
)
}
\arguments{
\item{...}{Parameter to GetAssayData.}
}
\description{
Pairwise DEG using limma-trend
# citation: https://ucdavis-bioinformatics-training.github.io/2019-single-cell-RNA-sequencing-Workshop-UCD_UCSF/scrnaseq_analysis/scRNA_Workshop-PART6.html
}
\examples{
DEG <- DEG_limma_trend(pbmc, 'Annotation', ident.1 = 'Memory CD4 T', ident.2 = 'Naive CD4 T')
# includes random effects
pbmc$batch <- c(rep('sample1',1000), rep('sample2',1000),rep('sample3',638))
DEG <- DEG_limma_trend(pbmc, 'Annotation', ident.1 = 'Memory CD4 T', ident.2 = 'Naive CD4 T', batch.colname = 'batch')
}
