% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{DEG_limma_voom}
\alias{DEG_limma_voom}
\title{Pairwise DEG using limma-voom}
\usage{
DEG_limma_voom(
  object,
  group.colname,
  ident.1,
  ident.2,
  batch.colname = NULL,
  max.cpm.cut.off = 1,
  duplicateCorrelation.name = NULL,
  ...
)
}
\arguments{
\item{...}{Parameter to GetAssayData.}
}
\description{
Pairwise DEG using limma-voom
}
\examples{
DEG <- DEG_limma_voom(pbmc, 'Annotation', ident.1 = 'Memory CD4 T', ident.2 = 'Naive CD4 T')
# includes random effects
pbmc$batch <- c(rep('sample1',1000), rep('sample2',1000),rep('sample3',638))
DEG <- DEG_limma_voom(pbmc, 'Annotation', ident.1 = 'Memory CD4 T', ident.2 = 'Naive CD4 T', batch.colname = 'batch')
}
