% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_MAST.R
\name{DEG_MAST}
\alias{DEG_MAST}
\title{Differential expression using MAST}
\usage{
DEG_MAST(
  object,
  group.colname,
  ident.1,
  ident.2,
  covariates = NULL,
  random_effect.vars = NULL,
  min.pct = 0.2,
  logfc.threshold = 0,
  method = "bayesglm",
  ebayes = T,
  pseudo_count = 1,
  parallel = T,
  n.cores = 4,
  p.adjust.method = "fdr",
  ...
)
}
\arguments{
\item{object}{A Seurat object.}

\item{ident.1}{One cluster in groupname.}

\item{ident.2}{Another cluster in groupname.}

\item{covariates}{Column names in meta.data for confounding effect correction.}

\item{random_effect.vars}{Covariates accounted for random effect.}

\item{method}{Argument to zlm. Character vector, either 'glm', 'glmer' or 'bayesglm'.}

\item{ebayes}{Argument to zlm. If TRUE, regularize variance using empirical bayes method.}

\item{n.cores}{How many cores to use.}

\item{groupname}{Column name of meta.data including 2 groups to performing differential expression.}

\item{min.cell}{The minimum number of cells in each group.}
}
\description{
Differential expression using MAST
}
\examples{
pbmc <- scPioneer::pbmc
pbmc$batch <- c(rep('sample1',1000), rep('sample2',1000),rep('sample3',638))
subs4 <- subset(pbmc, Annotation \%in\% c('CD14+ Mono','FCGR3A+ Mono'))
subs4 <- subs4[rowSums(GetAssayData(subs4, slot = 'counts')) > 100, ]
MASTdf <- DEG_MAST(subs4, 'Annotation', ident.1 = 'CD14+ Mono', ident.2 = 'FCGR3A+ Mono', covariates = 'percent.mt')
MASTdf <- DEG_MAST(subs4, 'Annotation', ident.1 = 'CD14+ Mono', ident.2 = 'FCGR3A+ Mono', covariates = 'percent.mt', random_effect.vars = 'batch', method = 'glmer', ebayes = F)

}
\references{
Seurat::MASTDETest; Andrew McDavid, Greg Finak and Masanao Yajima (2017). 
MAST: Model-based Analysis of Single Cell Transcriptomics. R package version 1.26.0. 
https://github.com/RGLab/MAST/. https://github.com/satijalab/seurat/issues/3712
}
