% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{DEG_MAST_RE}
\alias{DEG_MAST_RE}
\title{Pairwise DEG using MAST to account for random effect.}
\usage{
DEG_MAST_RE(
  object,
  group.colname,
  ident.1,
  ident.2,
  batch.colname = NULL,
  freq_expressed = 0.1,
  adjust.p.method = "fdr",
  Adaptive_thres = F
)
}
\description{
Firstly provide log2(x + 1) transformed values (either raw counts or TPMs)
directly to MAST without doing normalization (just transformation). 
The "ngeneson" parameter in the model will account for "cell size" and act as a sort of normalization. 
Refer to https://github.com/kdzimm/PseudoreplicationPaper/issues/2
}
\examples{
pbmc$batch <- c(rep('sample1',1000), rep('sample2',1000),rep('sample3',638))
subs4 <- subset(pbmc, Annotation \%in\% c('CD14+ Mono','FCGR3A+ Mono'))
MASTdf <- DEG_MAST_RE(subs4, 'Annotation', ident.1 = 'CD14+ Mono', ident.2 = 'FCGR3A+ Mono', batch.colname = 'batch',freq_expressed = 0.2)
MASTdf <- MASTdf[MASTdf$FDR < 0.05 & MASTdf$avg_logFC > 0.15,]
}
