% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_Meld.R
\name{DEG_Meld}
\alias{DEG_Meld}
\title{Pairwise DEG using Meld
Refer to https://nbviewer.org/github/KrishnaswamyLab/MELD/blob/main/notebooks/MELD_thresholding.Tcell.ipynb}
\usage{
DEG_Meld(
  object,
  groupname,
  ident.1,
  ident.2,
  covariates = NULL,
  test.Seurat = NULL,
  python2use = "/media/london_B/huangzh/software/miniconda3/envs/FR-Perturb/bin/python",
  meld_script_path = "./Meld_all.py",
  meta_script_path = "./Meld_meta.py",
  deg_script_path = "./Meld_deg.py",
  classes_script_path = "./Meld_classes.py",
  min.cell = 50,
  min.pct = 0.2,
  logfc.threshold = 0,
  on_disk = T,
  tmpdir = tempdir(),
  test.meld = c("rank", "t-test", "wald", "lrt"),
  p.adjust.method = c("bonferroni", "holm", "hochberg", "hommel", "BH", "BY", "fdr",
    "none"),
  return.classes = F
)
}
\arguments{
\item{test.Seurat}{Employ DE methods built in Seurat.}

\item{on_disk}{Whether to save temporary results on disk.}

\item{test.meld}{Employ DE methods built in Meld.}

\item{return.classes}{whether to return clusters of cells.}

\item{...}{Parameter to GetAssayData.}
}
\description{
Pairwise DEG using Meld
Refer to https://nbviewer.org/github/KrishnaswamyLab/MELD/blob/main/notebooks/MELD_thresholding.Tcell.ipynb
}
\examples{
pbmc <- readRDS('../data/pbmc_v4.rds')
pbmc <- subset(pbmc, Annotation \%in\% c('B','NK'))
pbmc$Annotation <- factor(pbmc$Annotation, levels = unique(pbmc$Annotation))
# includes random effects
pbmc$batch <- c(rep('sample1',100), rep('sample2',100),rep('sample3',299))
DEG <- DEG_Meld(pbmc, 'Annotation', 'B','NK')
DEG <- DEG_Meld(pbmc, 'Annotation', 'B','NK', test.Seurat = 'MAST', covariates = c('batch', 'percent.mt'))

}
