% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_miloDE.R
\name{DEG_miloDE}
\alias{DEG_miloDE}
\title{Pairwise DEG using miloDE. Employ edgeR testing (using glmQLFit) within each neighbourhood.
Refer to https://github.com/MarioniLab/miloDE?tab=readme-ov-file}
\usage{
DEG_miloDE(
  object,
  groupname,
  ident.1,
  ident.2,
  covariates = NULL,
  samplename = NULL,
  min.pct = 0.2,
  logfc.threshold = 0,
  reducedDim_name = "PCA",
  k = 25,
  order = 2,
  min_n_cells_per_sample = 3,
  min_count = 3,
  min.cell = 50,
  filtering = TRUE,
  n.cores = 1
)
}
\arguments{
\item{k}{defines how many neighbours to use for the neighbourhood assignment. Default k = 25.}

\item{order}{In c(1,2), defines which order of neighbours to use. Default order = 2.}

\item{...}{Parameter to GetAssayData.}
}
\description{
Pairwise DEG using miloDE. Employ edgeR testing (using glmQLFit) within each neighbourhood.
Refer to https://github.com/MarioniLab/miloDE?tab=readme-ov-file
}
\examples{
obj <- readRDS('./data/obj_pct01_logfc01.rds')
DEG <- DEG_miloDE(object = obj, samplename = 'Sample',groupname = 'crispr', min.cell = 20,covariates = c('crispr','Sample','percent.mt'))

}
