% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_mixscape.R
\name{DEG_mixscape}
\alias{DEG_mixscape}
\title{DEG using mixscape}
\usage{
DEG_mixscape(
  object,
  groupname,
  target_gene,
  nt.cell.class = "NT",
  covariates = NULL,
  ndims = NULL,
  num.neighbors = 20,
  assay = "RNA",
  new.assay.name = "PRTB",
  split.by = NULL,
  nt.class.name = "NT",
  min.de.genes = 5,
  iter.num = 10,
  de.assay = "RNA",
  test.use = "wilcox",
  min.pct = 0.1,
  logfc.threshold = 0.01
)
}
\description{
DEG using mixscape
}
\examples{
eccite <- LoadData(ds = "thp1.eccite")
DEG <- DEG_mixscape(eccite, groupname = 'gene', nt.class.name = 'NT',target_gene = 'STAT2' )

}
