% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_Monocle3.R
\name{DEG_Monocle3}
\alias{DEG_Monocle3}
\title{Pairwise DEG using Monocle3}
\usage{
DEG_Monocle3(
  object,
  groupname,
  ident.1 = ident.1,
  ident.2 = ident.2,
  batchname = NULL,
  min.pct = 0.2,
  logfc.threshold = 0,
  expression_family = c("quasipoisson", "negbinomial", "poisson", "binomial", "gaussian",
    "zipoisson", "zinegbinomial", "mixed-negbinomial"),
  adjust.method = c("bonferroni", "holm", "hochberg", "hommel", "BH", "BY", "fdr",
    "none"),
  n.cores = 8,
  ...
)
}
\arguments{
\item{groupname}{Colname including two groups for DEG analysis, e.g. SNP and WT labels.}

\item{batchname}{Colname including covariate in meta.data.}

\item{expression_family}{DEG models passed to fit_models().}

\item{...}{Parameter to GetAssayData.}
}
\description{
Pairwise DEG using Monocle3
}
\examples{
pbmc <- readRDS('../data/pbmc_v4.rds')
pbmc <- subset(pbmc, Annotation \%in\% c('B','NK'))
pbmc$Annotation <- factor(pbmc$Annotation, levels = unique(pbmc$Annotation))
# includes random effects
pbmc$batch <- c(rep('sample1',100), rep('sample2',100),rep('sample3',299))
DEG <- DEG_Monocle3(pbmc, groupname = 'Annotation',  batchname  = 'batch')

}
