% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_negbinom_Wald.R
\name{DEG_negbinom_Wald}
\alias{DEG_negbinom_Wald}
\title{Fitting gene expression using Negative Binomial Generalized linear model.}
\usage{
DEG_negbinom_Wald(
  object,
  groupname,
  ident.1,
  ident.2,
  p.adjust.method = "fdr",
  min.pct = 0.2,
  logfc.threshold = 0,
  covariates = NULL,
  n.cores = 4
)
}
\arguments{
\item{object}{A Seurat object.}

\item{groupname}{Column name of meta.data including 2 groups to performing differential expression.}

\item{ident.1}{One cluster in groupname.}

\item{ident.2}{Another cluster in groupname.}

\item{covariates}{Column names in meta.data for confounding effect correction.}

\item{min.cell}{The minimum number of cells in each group.}
}
\description{
Fitting gene expression using Negative Binomial Generalized linear model.
}
\examples{
pbmc <- scPioneer::pbmc
pbmc$batch <- c(rep('sample1',100), rep('sample2',100),rep('sample3',299))
coefdf <- DEG_negbinom_Wald(pbmc, 'Annotation', 'B','NK', n.cores = 8)
coefdf <- DEG_negbinom_Wald(pbmc, 'Annotation', 'B','NK', covariates = c('batch','percent.mt'), n.cores = 8)

}
\references{
https://stat.ethz.ch/R-manual/R-devel/library/MASS/html/glm.nb.html
}
