% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_poisson_Wald.R
\name{DEG_poisson_Wald}
\alias{DEG_poisson_Wald}
\title{Fitting gene expression using Poisson Regression model. 
To transform the non-linear relationship to linear form, a link function is used which is the log for Poisson Regression. 
For that reason, a Poisson Regression model is also called log-linear model. The general mathematical form of Poisson Regression model is:
log(y)=α+β1x1+β2x2+....+βnxn
Where,
y: Is the response variable
α and β: are numeric coefficients, α being the intercept, sometimes α also is represented by β0, it's the same
x is the predictor/explanatory variable.}
\usage{
DEG_poisson_Wald(
  object,
  groupname,
  ident.1,
  ident.2,
  p.adjust.method = "fdr",
  min.pct = 0.2,
  logfc.threshold = 0,
  covariates = NULL,
  n.cores = 4
)
}
\arguments{
\item{object}{A Seurat object.}

\item{groupname}{Column name of meta.data including 2 groups to performing differential expression.}

\item{ident.1}{One cluster in groupname.}

\item{ident.2}{Another cluster in groupname.}

\item{covariates}{Column names in meta.data for confounding effect correction.}

\item{min.cell}{The minimum number of cells in each group.}
}
\value{
A coefficient data frame. The first column named Estimate is the coefficient values of α (intercept), β1 and so on.
}
\description{
Fitting gene expression using Poisson Regression model. 
To transform the non-linear relationship to linear form, a link function is used which is the log for Poisson Regression. 
For that reason, a Poisson Regression model is also called log-linear model. The general mathematical form of Poisson Regression model is:
log(y)=α+β1x1+β2x2+....+βnxn
Where,
y: Is the response variable
α and β: are numeric coefficients, α being the intercept, sometimes α also is represented by β0, it's the same
x is the predictor/explanatory variable.
}
\examples{
pbmc <- scPioneer::pbmc
pbmc$batch <- c(rep('sample1',100), rep('sample2',100),rep('sample3',299))
coefdf <- DEG_poisson_Wald(pbmc, 'Annotation', 'B','NK', n.cores = 8)
coefdf <- DEG_poisson_Wald(pbmc, 'Annotation', 'B','NK', covariates = c('batch','percent.mt'), n.cores = 8)

}
\references{
https://www.bookdown.org/rwnahhas/RMPH/blr-likelihood.html;
 https://bookdown.org/ltupper/340f21_notes/glm-inference-tests.html#wald-tests;
 https://www.dataquest.io/blog/tutorial-poisson-regression-in-r/#:~:text=A%20Poisson%20Regression%20model%20is,form%20by%20some%20unknown%20parameters.
}
