% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_SCDE.R
\name{DEG_SCDE}
\alias{DEG_SCDE}
\title{Pairwise DEG using SCDE}
\usage{
DEG_SCDE(
  object,
  groupname,
  ident.1,
  ident.2 = NULL,
  batchname = NULL,
  min.pct = 0.2,
  logfc.threshold = 0,
  adjust.method = c("bonferroni", "holm", "hochberg", "hommel", "BH", "BY", "fdr",
    "none"),
  length.out = 400,
  n.randomizations = 100,
  n.cores = 1,
  verbose = 1,
  ...
)
}
\arguments{
\item{groupname}{Colname including two groups for DEG analysis, e.g. SNP and WT labels.}

\item{batchname}{Colname including covariate in meta.data.}

\item{length.out}{number of points (resolution) of the expression magnitude grid (default: 400). Note: larger numbers will linearly increase memory/CPU demands.}

\item{n.randomizations}{number of bootstrap randomizations to be performed.}

\item{...}{Parameter to GetAssayData.}

\item{min.lib.size}{Minimum number of genes detected in a cell. Cells with fewer genes will be removed (default: 1e3)}

\item{min.reads}{Minimum number of reads per gene. Genes with fewer reads will be removed (default: 1)}

\item{min.detected}{Minimum number of cells a gene must be seen in. Genes not seen in a sufficient number of cells will be removed (default: 1)}
}
\description{
Pairwise DEG using SCDE
}
\examples{
pbmc <- readRDS('../data/pbmc_v4.rds')
pbmc <- subset(pbmc, Annotation \%in\% c('B','NK'))
pbmc$Annotation <- factor(pbmc$Annotation, levels = unique(pbmc$Annotation))
# includes random effects
pbmc$batch <- c(rep('sample1',100), rep('sample2',100),rep('sample3',299))
DEG <- DEG_SCDE(pbmc, groupname = 'Annotation', ident.1 = 'B', ident.2 = 'NK', batchname  = 'batch', verbose = 0)

}
\references{
https://hms-dbmi.github.io/scde/index.html; https://hms-dbmi.github.io/scw/differential-expression.html
}
