% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_SCEPTRE.R
\name{DEG_SCEPTRE}
\alias{DEG_SCEPTRE}
\title{DEG using SCEPTRE}
\usage{
DEG_SCEPTRE(
  object,
  covariates,
  grna_targetdf,
  grna_matrix,
  moi = "high",
  to_ensembl = T,
  control_group = "default",
  p.adjust.method = "fdr",
  grna_method = "default",
  grna_threshold = 5,
  distance_threshold = 5e+06,
  p_mito_threshold = 0.075,
  parallel = TRUE,
  n_processors = "auto"
)
}
\arguments{
\item{distance_threshold}{distance_threshold bases of one another and 
on the same chromosome are included in the cis discovery set.}
}
\description{
DEG using SCEPTRE
}
\examples{
object <- readRDS('../data/obj_pct01_logfc01.rds')
metaf <- object@assays$RNA@meta.features
grna_targetdf <- object@assays$GDO@meta.features
# select sgRNA id with position info
grna_targetdf <- grna_targetdf[(!is.na(grna_targetdf$chr)) | grepl('non-targeting',grna_targetdf$grna_target),]
grna_targetdf <- grna_targetdf[, c('grna_id', 'grna_target', 'chr', 'start', 'end')]
grna_targetdf <- left_join(grna_targetdf, data.frame(grna_target = metaf$response_name,
                                                   response_id = metaf$response_id))
grna_targetdf$grna_target[!is.na(grna_targetdf$response_id)] <- grna_targetdf$response_id[!is.na(grna_targetdf$response_id)]
grna_targetdf <- grna_targetdf[,-6]
grna_matrix <- object@assays$GDO@counts[grna_targetdf$grna_id,]
result_high <- DEG_SCEPTRE(hmoi, covariates = 'Sample', grna_targetdf, grna_matrix, moi = 'high')
result_low <- DEG_SCEPTRE(hmoi, covariates = 'Sample', grna_targetdf, grna_matrix, moi = 'low', grna_threshold = 2)
}
