% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_scMAGeCK.R
\name{DEG_scMAGeCK}
\alias{DEG_scMAGeCK}
\title{DEG using scMAGeCK}
\usage{
DEG_scMAGeCK(
  object,
  assay_gdo = "GDO",
  assay_scripr = "scripr_thred",
  covariates = NULL,
  BARCODE = "/media/london_B/huangzh/project/benchmark_perturb/TEMP/BARCODE_10x.txt",
  module = c("RRA", "LR"),
  NEGCTRL = "non-targeting",
  target_gene = "CD52",
  label = "LR_result",
  PERMUTATION = 100,
  GENE_FRAC = 0.4
)
}
\arguments{
\item{module}{RRA: test the association with a known marker using scMAGeCK-RRA; LR module for large-scale association test.}

\item{GENE_FRAC}{A paramter for filtering low expressed genes. By default, only genes that have expressions in at least that fractions of cells (in raw count table) are kept. If raw count table is not found, will look into scaled expression, and only keep genes whose expression is greater than zero in at least that fraction of cells. Default: 0.01 description}
}
\description{
DEG using scMAGeCK
}
\examples{
hmoi <- readRDS('../data/obj_pct01_logfc01.rds')
result <- DEG_scMAGeCK(hmoi, covariates = 'Sample')

}
