% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_scvi.R
\name{DEG_scvi}
\alias{DEG_scvi}
\title{Pairwise DEG using scvi}
\usage{
DEG_scvi(
  object,
  groupname,
  ident.1,
  ident.2,
  covariates = NULL,
  n.cores = 8,
  min.pct = 0.2,
  logfc.threshold = 0,
  python2use =
    "/lustre/home/kwxiong/Huangzhihao/software/miniconda3/envs/scvi-env/bin/python",
  scvi_script_path =
    "/lustre/home/kwxiong/Huangzhihao/project/benchmark_perturb//shell/Perturb_benchmark-main/script/scvi.py"
)
}
\arguments{
\item{on_disk}{Whether to save temporary results on disk.}

\item{test.Seurat}{Employ DE methods built in Seurat.}

\item{test.meld}{Employ DE methods built in Meld.}

\item{return.classes}{whether to return clusters of cells.}

\item{...}{Parameter to GetAssayData.}
}
\description{
Pairwise DEG using scvi
}
\examples{
pbmc <- readRDS('/media/london_B/huangzh/project/benchmark_perturb/data/pbmc_v4.rds')
pbmc <- subset(pbmc, Annotation \%in\% c('B','NK'))
pbmc$Annotation <- factor(pbmc$Annotation, levels = unique(pbmc$Annotation))
# includes random effects
pbmc$batch <- c(rep('sample1',100), rep('sample2',100),rep('sample3',299))
DEG <- DEG_scvi(pbmc, 'Annotation', 'B','NK')
DEG <- DEG_scvi(pbmc, 'Annotation', 'B','NK', covariates = c('batch'))

}
\references{
https://docs.scvi-tools.org/en/stable/tutorials/notebooks/scrna/scVI_DE_worm.html
}
