% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_Seurat_DESeq2.R
\name{DEG_Seurat_DESeq2}
\alias{DEG_Seurat_DESeq2}
\title{DESeq2 method in Seurat FindMarkers}
\usage{
DEG_Seurat_DESeq2(
  object,
  ident.1,
  ident.2,
  group.by = NULL,
  min.pct = 0,
  logfc.threshold = 0,
  verbose = T,
  p.adjust.method = NULL,
  ...
)
}
\description{
DESeq2 method in Seurat FindMarkers
}
