% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_Seurat_poisson.R
\name{DEG_Seurat_poisson}
\alias{DEG_Seurat_poisson}
\title{Poisson method in Seurat FindMarkers}
\usage{
DEG_Seurat_poisson(
  object,
  ident.1,
  ident.2,
  group.by = NULL,
  min.pct = 0,
  logfc.threshold = 0,
  latent.vars = NULL,
  p.adjust.method = NULL,
  verbose = T,
  ...
)
}
\description{
Poisson method in Seurat FindMarkers
}
