% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_Seurat_t.R
\name{DEG_Seurat_t}
\alias{DEG_Seurat_t}
\title{t method in Seurat FindMarkers}
\usage{
DEG_Seurat_t(
  object,
  ident.1,
  ident.2,
  group.by = NULL,
  p.adjust.method = NULL,
  min.pct = 0,
  logfc.threshold = 0,
  verbose = T,
  ...
)
}
\description{
t method in Seurat FindMarkers
}
