% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_Seurat_wilcox.R
\name{DEG_Seurat_wilcox}
\alias{DEG_Seurat_wilcox}
\title{Wilcox method in Seurat FindMarkers}
\usage{
DEG_Seurat_wilcox(
  object,
  ident.1,
  ident.2,
  group.by = NULL,
  min.pct = 0,
  logfc.threshold = 0,
  p.adjust.method = NULL,
  verbose = T,
  ...
)
}
\description{
Wilcox method in Seurat FindMarkers
}
