% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimPlot_split.R
\name{DimPlot_split}
\alias{DimPlot_split}
\title{Split DimPlot with separated sample and all samples.}
\usage{
DimPlot_split(
  object,
  split.by = NULL,
  group.by = NULL,
  reduction.name = "umap",
  add.plot = c("ADD_ALL_PLOT"),
  add.plot.title = "All samples",
  add.location = 1,
  cols = NULL,
  ...
)
}
\description{
Split DimPlot with separated sample and all samples.
}
\examples{
pbmc_small$Sample <- c(rep('a', 20), rep('b',30), rep('c',30))
DimPlot_split(pbmc_small, split.by = 'Sample', add.plot.title = 'abc')
DimPlot_split(pbmc_small, split.by = 'Sample', add.plot.title = 'ab', add.plot = c('a','b'))
DimPlot_split(pbmc_small, split.by = 'Sample', add.plot.title = 'abc', add.location = 3)

}
