% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimPlot_split.R
\name{DimPlot_split2}
\alias{DimPlot_split2}
\title{Split DimPlot with separated sample and all samples.}
\usage{
DimPlot_split2(
  object,
  split.by = NULL,
  group.by = NULL,
  reduction.name = "umap",
  x = "umap_1",
  y = "umap_2",
  add.plot = c("ADD_ALL_PLOT"),
  pt.size = 0.1,
  add.plot.title = "All samples",
  add.location = 1,
  color.use = NULL,
  panel.spacing = 0.5,
  strip.text.switch = NULL,
  strip.text.size = 12,
  strip.text.angle = NULL,
  strip.text.hjust = NULL,
  strip.text.vjust = NULL,
  remove.axis.x = F,
  remove.axis.y = F,
  legend.position = "right",
  order.cell = T,
  ...
)
}
\description{
Split DimPlot with separated sample and all samples.
}
\examples{
pbmc_small$Sample <- c(rep('a', 20), rep('b',30), rep('c',30))
DimPlot_split2(pbmc_small, split.by = 'Sample', add.plot.title = 'abc')
DimPlot_split2(pbmc_small, split.by = 'Sample', add.plot.title = 'ab', add.plot = c('a','b'))
DimPlot_split2(pbmc_small, split.by = 'Sample', add.plot.title = 'abc', add.location = 3)

}
