% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion_statistics.R
\name{diver_t_test}
\alias{diver_t_test}
\title{Calculate diversity of cell types. used Simpson’s index to characterize the composition of cell types.}
\usage{
diver_t_test(
  meta,
  annotation = "Annotation",
  sample = "Sample",
  group = "Group",
  index = c("simpson", "invsimpson", "shannon"),
  alternative = "two.sided",
  adjust.method = c("BH", "holm", "hochberg", "hommel", "bonferroni", "BY", "fdr",
    "none"),
  plabel = c("p.adj.signif", "p.ad"),
  add = "jitter",
  xlab = NULL,
  ylab = NULL,
  x.angle = 45,
  x.hjust = 1,
  x.vjust = 1,
  melt.data = T,
  return.all = F,
  return.plot = F
)
}
\description{
Calculate diversity of cell types. used Simpson’s index to characterize the composition of cell types.
}
\examples{
###Calculate diversity for each sample.
library(rstatix)
library(ggpubr)
diver_t_test(meta200, annotation = 'maingroups',sample = 'Sample',
group = 'Group', index = 'simpson',return.plot = T)

}
