% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dot.R
\name{DotPlot2}
\alias{DotPlot2}
\title{DotPlot modified from Seurat::DotPlot}
\usage{
DotPlot2(
  object,
  features,
  assay = NULL,
  col.min = -2.5,
  col.max = 2.5,
  dot.min = 0,
  dot.scale = 6,
  color = scPalette_heatmap_bar(),
  group.by = NULL,
  split.by = NULL,
  scale.by = "radius",
  scale.min = NA,
  scale.max = NA,
  y.size = 12,
  x.size = 12,
  x.vjust = 0.5,
  scale = T,
  x.hjust = 1,
  x.angle = 90,
  legend.color.name = "Average\\nExpression",
  cols = c("lightgrey", "blue")
)
}
\description{
DotPlot modified from Seurat::DotPlot
}
