% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_angrycell_result.R
\name{DotPlot_anno}
\alias{DotPlot_anno}
\title{Dotplot for marker gene from the result of angrycell.}
\usage{
DotPlot_anno(
  object,
  features = NULL,
  angrydf = NULL,
  assay = NULL,
  auto.order = T,
  sort.name.per.group.byAnnot = F,
  human.to.mouse = F,
  col.min = -2.5,
  col.max = 2.5,
  dot.min = 0,
  dot.scale = 6,
  color = scPalette_heatmap_bar(),
  group.by = NULL,
  split.by = NULL,
  scale.by = "radius",
  scale.min = NA,
  scale.max = NA,
  y.size = 8,
  x.size = 8,
  x.vjust = 0.5,
  x.hjust = 1,
  x.angle = 90,
  coord.flip = T,
  legend.color.name = "Average\\nExpression",
  cols = c("lightgrey", "blue"),
  order.method = "DEG"
)
}
\description{
Dotplot for marker gene from the result of angrycell.
}
\examples{
angrydf <- angrycell(pbmc, select.db = c('db_user'), show.top = 2)
DotPlot_anno(pbmc, angrydf = angrydf)

}
