% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dot.R
\name{dotplot_g}
\alias{dotplot_g}
\title{dotplot group}
\usage{
dotplot_g(
  plotdata,
  highlight = NULL,
  cell.order = NULL,
  gene.order = NULL,
  text.size = 8,
  text.x.size = 8,
  text.y.size = 8,
  legend.title.size = 15,
  legend.text.size = 12,
  color.use = rev(brewer.pal(11, "RdYlBu")),
  facet.toward = "col",
  scale.min = NA,
  scale.max = NA,
  scale.size = c(0, 1.5, 3),
  x_position = "top",
  y_position = "left",
  legend.position = "right",
  strip.text.size = 8,
  strip.angle = 45,
  strip.vjust = 0,
  strip.hjust = 0,
  axis.x.position = NULL,
  axis.y.position = NULL,
  strip.text.switch = NULL,
  panel.spacing = 0.5,
  clip = "on",
  max.pair = 50
)
}
\description{
dotplot group
}
\examples{
ov_PT <- subset(ov_obj, Group_abb == 'PT')
ov_MT <- subset(ov_obj, Group_abb == 'MT')
datalist_PT <- RunCellphoneDB(object = ov_PT,annotation = 'celltype',outdir = './test/PT')
datalist_MT <- RunCellphoneDB(object = ov_MT,annotation = 'celltype',outdir = './test/MT')
#file_mean <- read.table(file='./test/PT/means.txt', header=T, sep='\t', check.names=F, as.is=T,stringsAsFactors = F)
#file_pval <- read.table(file='./test/PT/pvalues.txt', header=T, sep='\t', check.names=F, as.is=T,stringsAsFactors = F)
#datalist_PT <- list(mean = file_mean, pval = file_pval)
#file_mean <- read.table(file='./test/MT/means.txt', header=T, sep='\t', check.names=F, as.is=T,stringsAsFactors = F)
#file_pval <- read.table(file='./test/MT/pvalues.txt', header=T, sep='\t', check.names=F, as.is=T,stringsAsFactors = F)
#datalist_MT <- list(mean = file_mean, pval = file_pval)
df_PT <- interaction_data(datalist_PT$mean, datalist_PT$pval,cellchat.evidence = T)
df_MT <- interaction_data(datalist_MT$mean, datalist_MT$pval,cellchat.evidence = T)
df_PT$organ <- 'PT'
df_MT$organ <- 'MT'
df <- as.data.frame(rbind(df_PT, df_MT))
chemo_pathways <- 'CCL|CXCL|XCR|CX3C'
df_DC <- df[df$ligand == 'DC'& grepl(chemo_pathways, df$pathway_name),]
dotplot_g(df_DC, strip.angle = 90)
gridlist <-dotplot_g(df_DC, clip = 'off', panel.spacing = 0, x_position = 'bottom')
grid::grid.draw(gridlist)

}
