% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dot.R
\name{dotplot_multicolor}
\alias{dotplot_multicolor}
\title{hightlight dotplot}
\usage{
dotplot_multicolor(
  exp,
  ident,
  highlight,
  cell.order = NULL,
  gene.order = NULL,
  circle.size.as = c("expression percent", "expressed cell number"),
  text.size = 8,
  text.x.size = 8,
  text.y.size = 8,
  legend.title.size = 8,
  legend.text.size = 8,
  do.scale.circle.size = T,
  legend.key.size = 1,
  legend.box = "horizonal",
  coord_flip = T,
  color.use = c("grey", "yellow", "red"),
  noSig.circle.color = "white",
  gene_text_color = NULL,
  scale.min = NA,
  scale.max = NA,
  x_position = "bottom",
  y_position = "left",
  text.x.angle = 90,
  x.vjust = 1,
  x.hjust = 1,
  text.y.angle = 0,
  y.vjust = 0.5,
  y.hjust = 0.5,
  legend.title.fill = "Average\\nExpression",
  legend.title.percent = "Percent of\\nExpression",
  legend.position = "right",
  line.size = 0.5,
  circle.border.size = 1.2,
  plot.margin = unit(c(0.1, 0.1, 0.1, 0.1), "cm")
)
}
\description{
hightlight dotplot
}
