% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dot.R
\name{dotplot_Rscript}
\alias{dotplot_Rscript}
\title{DotPlot of all cell pairs.}
\usage{
dotplot_Rscript(
  mfilePath,
  pfilePath,
  outdir,
  Rlib =
    "/BGFS1/home/huangzh/workspace/project/single_cell_pipeline_development/BiosPipe_V0.1.0/lib/Rlib/BiosPipe_V0.1.0/",
  dotplot_func =
    "/BGFS1/home/huangzh/workspace/project/single_cell_pipeline_development/BiosPipe_V0.1.0/biospipe/modules/cellphoneDB/cellphoneDB_dotplot.R"
)
}
\arguments{
\item{mfilePath}{File path to file mean.txt from result of cellphoneDB.}

\item{pfilePath}{File path to file pvalues.txt form result of cellphoneDB.}

\item{outdir}{Path to save dotplot PDF.}

\item{Rlib}{R labrary path which was needed by dotplot.}

\item{dotplot_func}{R script of dotplot.}
}
\description{
DotPlot of all cell pairs.
}
\examples{
result_dir <- '/BGFS1/home/huangzh/workspace/project/r_package_development/biospiper/biospiper_V1.1.0/biospiper/test/'
dotplot_Rscript(mfilePath = paste0(result_dir, 'means.txt'), pfilePath = paste0(result_dir, 'pvalues.txt'),outdir = result_dir)
#dotplot PDF were saved in outdir/cellphoneDB/dotplot/.

}
