% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_enrichment.R
\name{enrichMSigDB}
\alias{enrichMSigDB}
\title{Function for enrichment of MSigDB}
\usage{
enrichMSigDB(
  markerlist = NULL,
  DEGfile = NULL,
  N,
  species = c("Homosapiens", "mm10")[1],
  keyType = c("SYMBOL", "ENTREZID", "ENSEMBL")[1],
  use.top.marker = 100,
  filter.gene.pattern = NULL,
  ont = "c2.cgp.v7.1",
  pvalueCutoff = 0.05,
  qvalueCutoff = 0.2
)
}
\arguments{
\item{markerlist}{Vector. Candidate different expression gene set for enrichment.}

\item{DEGfile}{Dataframe. DEG file result from FindAllMarker function in Seurat.}

\item{N}{number of set with all the genes to consider in the analysis.}

\item{species}{Select one species, Homosapiens or mm10.}

\item{keyType}{gene ID format.}

\item{use.top.marker}{Select top n marker genes for each cluster in DEG file.}

\item{filter.gene.pattern}{Filtering genes with some pattern, such as 'MT-'.}

\item{ont}{Select a database. Using listMSigDBnames() to list all databases' names.}

\item{pvalueCutoff}{P value cut-off.}

\item{qvalueCutoff}{Adjust P value. Only BH method in this version.}
}
\description{
Function for enrichment of MSigDB
}
