% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_scatter.R
\name{FeaturePlot2}
\alias{FeaturePlot2}
\title{Modified FeaturePlot, to reduce legend size, and add more colors.}
\usage{
FeaturePlot2(
  object,
  features,
  color = NULL,
  pt.size = 1,
  legend.position = c(1, 0.2),
  legend.key.size = 0.3,
  legend.text.size = 6,
  ncols = NULL,
  split.by = NULL,
  plot.margin = c(0, 1, 0, 0),
  AI.friendly = F,
  dpi = 300,
  AI.script =
    "/BGFS1/projectdata/jasper/project/ovarian_cancer/shell/functions/ggAIplot.R",
  ...
)
}
\arguments{
\item{legend.position}{Vector with 2 elements. Adjust position of legend by location of x/y axis.}
}
\description{
Modified FeaturePlot, to reduce legend size, and add more colors.
}
\examples{
FeaturePlot2(object = pbmc, features = c('CD3D','CD79A','CD14','FCGR3A'))
FeaturePlot2(object = pbmc, features = c('CD3D','CD79A','CD14','FCGR3A'),
legend.position = c(0.9,0.2))

}
