% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_gene}
\alias{filter_gene}
\title{Filter genes by min.pct and logfc.threshold}
\usage{
filter_gene(
  object,
  group.by = NULL,
  ident.1,
  ident.2 = NULL,
  min.pct = 0.01,
  logfc.threshold = 0,
  assay = "RNA",
  slot = "counts",
  return.fcresult = F
)
}
\description{
Filter genes by min.pct and logfc.threshold
}
\references{
Seurat::FoldChange
}
