% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_similarity.R
\name{Find_similar}
\alias{Find_similar}
\title{Detect similarity of different cell types using variable genes from integration data.}
\usage{
Find_similar(
  integrated.object = NULL,
  Annotation = "Annotation",
  var.genes.method = c("var", "pca"),
  pca.dims = NULL,
  pca.top.var = 50,
  assay = "integrated",
  clustering.method = "complete",
  clustering_distance = c("cormat", "pearson"),
  clustering_distance_cols = c("correlation", "euclidean", "binary"),
  clustering_distance_rows = c("correlation", "euclidean", "binary"),
  reverse.cor = T,
  k_clusters = 4,
  plot = T,
  return.data = F,
  layout = c("circular.dendrogram", "horizontal.dendrogram", "heatmap_cc"),
  circle.text.size = 2.2,
  circle.line.size = 1,
  circle.margin = 3,
  label.color = NULL,
  dend.color = NULL,
  heatmap.display.numbers = T
)
}
\arguments{
\item{var.genes.method}{Methods for high variable genes selection. var: var.genes from Seurat::FindVariableFeatures; pca: genes with higher SD in PCA dimensions.}

\item{clustering_distance}{Methods for measuring distance in matrix. cormat: construct cell x cell correlation matrix for pheatmap; pearson: distance defined as (1-Pearson correlation coefficient)/2.}

\item{clustering_distance_cols}{Methods for measuring distance of cells. Pass parameter to dist() function. binary is also called Jaccard.}

\item{clustering_distance_rows}{similar to clustering_distance_cols.}

\item{reverse.cor}{Calculate distance for hierarchical clustering. Refer to https://doi.org/10.1016/j.cell.2021.01.010.}
}
\description{
Detect similarity of different cell types using variable genes from integration data.
}
\examples{
myeloid.integrated <- readRDS('path_to_your_seurat_integrated.rds')
datalist <- Find_similar(integrated.object = myeloid.integrated, assay = 'integrated',
                        var.genes.method = 'var', Annotation = 'Annotation', plot = F, return.data = T)
#correlation heatmap
Find_similar(integrated.object = myeloid.integrated, assay = 'integrated',layout = 'heatmap',
var.genes.method = 'var', Annotation = 'Annotation', plot = T, return.data = F)
#dendrogram
Find_similar(integrated.object = myeloid.integrated, assay = 'integrated',layout = 'horizontal.dendrogram',
var.genes.method = 'var', Annotation = 'Annotation', k_clusters = 4, plot = T, return.data = F)
# cirlize denrogram
Find_similar(integrated.object = myeloid.integrated, assay = 'integrated',layout = 'circular.dendrogram',
var.genes.method = 'var', Annotation = 'Annotation', k_clusters = 4, plot = T, return.data = F)
###The distance defined as (1-Pearson correlation coefficient)/2.
Find_similar(integrated.object = myeloid.integrated, assay = 'integrated',layout = 'horizontal.dendrogram',var.genes.method = 'var',clustering_distance = 'pearson', Annotation = 'Annotation', plot = T, return.data = F)

}
