% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_process.R
\name{findpeak_boo}
\alias{findpeak_boo}
\title{Estimate  two side cut-off values of input-vector data.}
\usage{
findpeak_boo(v, frac = 0.1, adjust = 1)
}
\arguments{
\item{rawlist}{A list contains multi vectors in which elements have names.}
}
\value{
A list.
}
\description{
Estimate  two side cut-off values of input-vector data.
}
\examples{
data("pbmc")
Gene_counts <- pbmc@meta.data$nFeature_RNA
a <- density(Gene_counts,adjust = 1)
genecf <- findpeak_boo(Gene_counts)
gene_counts_cut_off = c(genecf$vp_cutoff,genecf$pv_cutoff)
print('gene count cut off: ')
print(gene_counts_cut_off)
#[1]  167 4381
hist(Gene_counts,prob=TRUE,breaks=200)
lines(density(Gene_counts,adjust = 0.5), col="blue", lwd=2) # add a density estimate with defaults
abline(v=gene_counts_cut_off[1],col='red')
abline(v=gene_counts_cut_off[2],col='red')
}
