% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_process.R
\name{FindVariableFeatures1}
\alias{FindVariableFeatures1}
\title{Find variable genes}
\usage{
FindVariableFeatures1(
  object,
  selection.method = c("vst", "mean.var.plot", "dispersion")[1],
  loess.span = 0.3,
  clip.max = "auto",
  num.bin = 20,
  binning.method = c("equal_width", "equal_frequency")[1],
  nfeatures = 2000,
  mean.cutoff = c(0.1, 8),
  dispersion.cutoff = c(1, Inf),
  assay = "RNA"
)
}
\description{
Find variable genes
}
