% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{FoldChange.Assay}
\alias{FoldChange.Assay}
\title{Unknown function}
\usage{
FoldChange.Assay(
  object,
  annotation,
  ident.1,
  ident.2,
  features = NULL,
  slot = "data",
  pseudocount.use = 1,
  fc.name = NULL,
  base = 2,
  ...
)
}
\description{
Unknown function
}
