% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{FoldChange.default}
\alias{FoldChange.default}
\title{Unknown function}
\usage{
FoldChange.default(
  object,
  cells.1,
  cells.2,
  mean.fxn,
  fc.name,
  features = NULL
)
}
\description{
Unknown function
}
