% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_scatter.R
\name{fplot1}
\alias{fplot1}
\title{Scater plot of gene expression.}
\usage{
fplot1(
  xy,
  exprmat,
  ncol = NULL,
  legend.title = "auto",
  pt.size = 0.1,
  color.use = c("grey", "yellow", "red"),
  order.expression = "ascending",
  AI.friendly = F,
  AI.script =
    "/BGFS1/projectdata/jasper/project/ovarian_cancer/shell/functions/ggAIplot.R"
)
}
\arguments{
\item{xy}{Cell location from seurat_object@reduction$tsne@cell.embeddings.}

\item{exprmat}{A normalize selected genes expression dataframe(cell x gene).}
}
\value{
A ggplot object.
}
\description{
Scater plot of gene expression.
}
\examples{
data('pbmc')
marker_to_plot <- c("MS4A1", "GNLY", "CD3E", "CD14", "FCER1A", "FCGR3A", "LYZ", "PPBP", "CD8A")
exp <- log2(as.data.frame(t(as.matrix(pbmc@assays$RNA@data[marker_to_plot,])))+1)
xy <- pbmc@reductions$umap@cell.embeddings
fplot1(xy,exp)

}
