% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_scatter.R
\name{fplot2}
\alias{fplot2}
\title{Scater plot of gene expression.}
\usage{
fplot2(
  object,
  features,
  reduction = "umap",
  ncol = NULL,
  assay = "RNA",
  label.cluster = NULL,
  label.size = 4,
  fontface = "plain",
  AI.friendly = F,
  AI.script =
    "/BGFS1/projectdata/jasper/project/ovarian_cancer/shell/functions/ggAIplot.R",
  legend.title = "auto",
  pt.size = 0.1,
  color.use = c("grey", "yellow", "red"),
  order.expression = NULL
)
}
\value{
A ggplot object.
}
\description{
Scater plot of gene expression.
}
\examples{
marker_to_plot <- c("MS4A1", "GNLY", "CD3E", "CD14", "FCER1A", "FCGR3A", "LYZ", "PPBP", "CD8A")
fplot2(pbmc,marker_to_plot)

}
