% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_scatter.R
\name{fplot3}
\alias{fplot3}
\title{Scater plot of gene expression.}
\usage{
fplot3(
  object,
  features,
  assay = "RNA",
  color = NULL,
  pt.size = 0.1,
  label.cluster = NULL,
  label.size = 3,
  order.expression = F,
  reduction = "umap",
  legend.title = "",
  legend.position = c(1, 0.2),
  legend.key.size = 0.3,
  legend.text.size = 6,
  strip.text.size = 13,
  plot.margin = c(0, 1, 0, 0),
  ncol = NULL,
  AI.friendly = F,
  AI.script =
    "/BGFS1/projectdata/jasper/project/ovarian_cancer/shell/functions/ggAIplot.R"
)
}
\value{
A ggplot object.
}
\description{
One gene one bar.
}
\examples{
marker_to_plot <- c("MS4A1", "GNLY", "CD3E", "CD14", "FCER1A", "FCGR3A", "LYZ", "PPBP", "CD8A")
fplot3(pbmc,marker_to_plot)

}
