% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_scatter.R
\name{fplot_label}
\alias{fplot_label}
\title{Scater plot of gene expression with label.}
\usage{
fplot_label(
  xy,
  exprmat,
  label.cluster,
  ncol = NULL,
  legend.title = "auto",
  reduction = "tsne",
  label.size = 4,
  fontface = "plain",
  pt.size = 0.1,
  AI.friendly = F,
  AI.script =
    "/BGFS1/projectdata/jasper/project/ovarian_cancer/shell/functions/ggAIplot.R",
  color.use = c("grey", "yellow", "red"),
  order.expression = "ascending"
)
}
\arguments{
\item{xy}{Cell location from seurat_object@reduction$tsne@cell.embeddings.}

\item{exprmat}{A normalize selected genes expression dataframe(cell x gene).}

\item{label.cluster}{A vector stored cluster labels in meta.data.}
}
\value{
A ggplot object.
}
\description{
Scater plot of gene expression with label.
}
\examples{
data('pbmc')
marker_to_plot <- c("MS4A1", "GNLY", "CD3E", "CD14", "FCER1A", "FCGR3A", "LYZ", "PPBP", "CD8A")
exp <- log2(as.data.frame(t(as.matrix(pbmc@assays$RNA@data[marker_to_plot,])))+1)
xy <- pbmc@reductions$umap@cell.embeddings
label.cluster <- pbmc$Annotation
fplot_label(xy,exp,label.cluster)

}
