% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{get_cutoff}
\alias{get_cutoff}
\title{Find the cutoff value of a curve.Default uses f(x) = ax^2 + bx +c to fit the curve.}
\usage{
get_cutoff(x, y, area.frac, model = "line3P", ...)
}
\arguments{
\item{x}{x axis of the curve.}

\item{y}{y axis of the curve.}

\item{area.frac}{Fraction of area under the curve to find}

\item{...}{other parameters in function trendline.
the intersection point on the curve.}
}
\value{
The upper and lower cutoff values
}
\description{
Find the cutoff value of a curve.Default uses f(x) = ax^2 + bx +c to fit the curve.
}
