% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{get_cutoff_NB}
\alias{get_cutoff_NB}
\title{get zero-fraction cut off value from negative binomial model.
Cluster with large phi is supposed to contain multi-cell types,
and cell marker genes have high zero fraction.}
\usage{
get_cutoff_NB(object, active.assay, celltype, initial.zero.frac)
}
\description{
get zero-fraction cut off value from negative binomial model.
Cluster with large phi is supposed to contain multi-cell types,
and cell marker genes have high zero fraction.
}
\examples{
get_cutoff_NB(pbmc, 'RNA', 'CD8 T',0.3)
}
