% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{get_HKcutoff}
\alias{get_HKcutoff}
\title{Using housekeeeping genes and negative binomial model to get the cutoff value
of dropout rate.}
\usage{
get_HKcutoff(object, active.assay, celltype, HKgenes, min.pct)
}
\arguments{
\item{object}{A seurat object.}

\item{active.assay}{Default assay in seurat object which used to model.}

\item{celltype}{The cell cluster name.}

\item{HKgenes}{House Keeping genes to model.}

\item{min.pct}{cell Fraction of no-zero expression.
the intersection point on the curve.}
}
\value{
gene list
}
\description{
Using housekeeeping genes and negative binomial model to get the cutoff value
of dropout rate.
}
