% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue_similar.R
\name{get_neareast_tissue}
\alias{get_neareast_tissue}
\title{To quantify the cell similarities among different tissues.}
\usage{
get_neareast_tissue(
  object,
  tissue_colnames,
  query_group,
  use.exist.reduction = NULL,
  assay = "RNA",
  npcs = 50,
  hvg.method = "vst",
  hvg.nfeatures = 2000,
  vars.to.regress = NULL,
  k = 30,
  seed = 123,
  n.perm = 1000,
  pval.cutoff = 0.05,
  n.core = 4
)
}
\arguments{
\item{object}{Seurat object.}

\item{tissue_colnames}{The colnumn name of seurat meta data store tissue labels.}

\item{query_group}{Which label is query tissue.}

\item{assay}{Surat assays.}

\item{npcs}{Parameter in Seurat::RunPCA.}

\item{hvg.method}{Select method for Seurat::FindVariableFeatures.}

\item{hvg.nfeatures}{Number of HVG.}

\item{vars.to.regress}{parameter in Seurat::ScaleData}

\item{k}{Number of nearest neighbor of each query cell.}

\item{seed}{Random seed for permutation test.}

\item{n.perm}{times of permutation test.}

\item{pval.cutoff}{P value cut off.}

\item{n.core}{Number of processes are applied when running permutation test.}
}
\value{
A binary matrix. Columns is tissues, rows is query cells. 
1 indicate connected.
}
\description{
Query for each cell's nearest neighbor among cells of all other tissues 
in the low-dimensional space. Inspired by Fig2A in 
https://doi.org/10.1016/j.cell.2019.10.003.
}
\examples{
main <- readRDS("/BGFS1/projectdata/project_xinhua_wangxipeng/result/rds_HGSOC/HGSOC_main_rm_cellcycle.rds")
net_mat <- get_neareast_tissue(object = main,tissue_colnames = 'Group_abb',
query_group = 'AS', vars.to.regress = 'percent.mt', n.core = 48)

}
