% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_process.R
\name{get_numPCs_CUM}
\alias{get_numPCs_CUM}
\title{Auto-select numbers of PCs and resolution}
\usage{
get_numPCs_CUM(
  object,
  do.plot = F,
  ndims = 100,
  pc.perc.low = 80,
  pc.perc.up = 95,
  gap.value = 0.001,
  return.pc.r = F
)
}
\arguments{
\item{object}{A seurat object.}

\item{ndims}{Total Number of PCs to compute and store (100 by default)}

\item{pc.perc.low}{The minimum value of accumulative PCstdev.}

\item{pc.perc.up}{The maximum value of accumulative PCstdev}

\item{gap.value}{the value of gap between two adjacent PCstdev.}

\item{return.pc.r}{Return PCs and resolution instead of the range of PCs.}

\item{plotfile}{The filepath of ElbowPlot file.}
}
\value{
A vector which contains lower PCs and upper PCs.
}
\description{
Find cut off value by top 80% - 95% accumulative PCstdev with gaps > 0.001.
}
\examples{
get_numPCs_CUM(pbmc, plotfile=NULL, ndims = 100, return.pc.r = T)
get_numPCs_CUM(pbmc)
}
