% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_process.R
\name{get_numPCs_JA}
\alias{get_numPCs_JA}
\title{Auto-select numbers of PCs and resolution.}
\usage{
get_numPCs_JA(object, dim = 100, p.val = 0.001, num.replicate = 100)
}
\arguments{
\item{object}{A seurat object.}

\item{p.val}{The cut off p.value from JackStraw.}

\item{plotfile}{the filepath of ElbowPlot file.}

\item{ndims}{Total Number of PCs to compute and store (100 by default)}
}
\value{
A vector which contains predicted PCs and resolution.
}
\description{

}
\examples{
get_numPCs_JA(pbmc,dim=10)
}
