% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{get_phi}
\alias{get_phi}
\title{get phi parameter of negative binomial model}
\usage{
get_phi(object, active.assay, celltype)
}
\description{
get phi parameter of negative binomial model
}
\examples{
get_phi(pbmc, 'RNA', 'Naive CD4 T')
}
