% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_symbol_ensembl_ID.R
\name{Get_symbol_ensembl_ID}
\alias{Get_symbol_ensembl_ID}
\title{Get genelist with symbol and ensembl ID.}
\usage{
Get_symbol_ensembl_ID(
  version = 98,
  species = "Human",
  using.local.file = F,
  file.dir = NULL,
  using.local.data = F,
  remove.duplicated = T,
  only.one.to.one = F
)
}
\arguments{
\item{version}{Version of ensembl.}

\item{using.local.file}{Whether to use annotation file from local directory instead of connecting to internet.}

\item{file.dir}{Directory for annotation file.}

\item{remove.duplicated}{Whether to remove duplicated gene symbol.}

\item{only.one.to.one}{Whether to keep rows in which both gene symbol and ensembl ID are unique.}
}
\description{
Get genelist with symbol and ensembl ID.
}
\examples{
df <- Get_symbol_ensembl_ID(version = 98)
df <- Get_orthologs_mouse_human(version = 98, using.local.data = T) # if cannot access biomaRt.

}
