% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{get_terms_P1}
\alias{get_terms_P1}
\title{Calculate p value of hypergeometric distribution test.}
\usage{
get_terms_P1(candidate, N, db)
}
\arguments{
\item{candidate}{Genes to test.}

\item{N}{Number of cells.}

\item{db}{Marker genes database for usage.}
}
\value{
Result of GO enrichment.
}
\description{
Calculate p value of hypergeometric distribution test.
}
