% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_enrichment.R
\name{get_terms_P_MsigDB}
\alias{get_terms_P_MsigDB}
\title{Hypergeometirc test by phyper function}
\usage{
get_terms_P_MsigDB(candidate, N, dbname)
}
\arguments{
\item{N}{number of set with all the genes to consider in the analysis.}
}
\description{
Hypergeometirc test by phyper function
}
