% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggAIplot.R
\name{ggAIplot.grid}
\alias{ggAIplot.grid}
\title{Raster points in UMAP in order to make umap pdf easily be loaded by AI. 
This function is suitable for multi-facet plots.}
\usage{
ggAIplot.grid(
  plot,
  facet.by,
  calibration = c(0, 0, 0, 0),
  width = 10,
  height = 10,
  dpi = 300
)
}
\description{
Raster points in UMAP in order to make umap pdf easily be loaded by AI. 
This function is suitable for multi-facet plots.
}
\author{
Li Ziyi
}
