% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pathway_result.R
\name{GO_all_plot}
\alias{GO_all_plot}
\title{GO bar plot}
\usage{
GO_all_plot(
  GOmat,
  color.use = NULL,
  show.num = 15,
  p.adjust.cutoff = 0.05,
  text.gap = 0.1,
  top.p.adjust = 1e-30,
  title = NULL
)
}
\description{
GO bar plot
}
