% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pathway_result.R
\name{GO_pair_plot}
\alias{GO_pair_plot}
\title{pairs barplot showing the result of GO.}
\usage{
GO_pair_plot(
  GOmat,
  show.num = 15,
  pvalue.adj = 0.05,
  text.gap = 0.1,
  top.p.adjust = 1e-30,
  text.max.length = 50,
  bar.width = 0.8,
  text.lineheight = 0.6,
  text.size = 3,
  show.gene.on.bar = F,
  gene.max.length = 10,
  gene.size = 3,
  gene.gap = 0.1,
  color = NULL,
  plot.title = NULL,
  legend.title = NULL
)
}
\description{
pairs barplot showing the result of GO.
}
