% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pathway_result.R
\name{GOheatmap}
\alias{GOheatmap}
\title{Heatmap of GO enrichment.}
\usage{
GOheatmap(
  GOmat,
  show.top.term = 5,
  log10.adj.P = F,
  na.as = 0.05,
  cell.order = NULL,
  term.order = NULL,
  coord_flip = F,
  color.use = NULL,
  grid.color = "black",
  text.y.size = 8,
  text.x.size = 10,
  legend.title.size = 10,
  legend.text.size = 10,
  max.length = 35,
  method = c("pheatmap", "ggplot2")[2],
  cluster_rows = FALSE,
  cluster_cols = FALSE
)

GOheatmap(
  GOmat,
  show.top.term = 5,
  log10.adj.P = F,
  na.as = 0.05,
  cell.order = NULL,
  term.order = NULL,
  coord_flip = F,
  color.use = NULL,
  grid.color = "black",
  text.y.size = 8,
  text.x.size = 10,
  legend.title.size = 10,
  legend.text.size = 10,
  max.length = 35,
  method = c("pheatmap", "ggplot2")[2],
  cluster_rows = FALSE,
  cluster_cols = FALSE
)
}
\arguments{
\item{GOmat}{Result of GO enrichment from clusterGO.}

\item{show.top.term}{Number of top GO terms of each cluster showed in dotplot.}

\item{cell.order}{Cell order showed in dotplot.}

\item{term.order}{GO terms order showed in dotplot.}

\item{coord_flip}{Parameter in ggplot.}

\item{color.use}{Color patterns used for showing p value.}

\item{text.y.size}{Text size of GO terms.}

\item{text.x.size}{Text size of Celltypes.}

\item{legend.title.size}{Text size of legend title.}

\item{legend.text.size}{Text size of legend text.}

\item{log.adj.P}{Show -log(adjust p value) instead of adjust p value.}
}
\description{
Heatmap of GO enrichment.

Heatmap of GO enrichment.
}
\examples{
markerlist <- FindAllMarkers(pbmc,only.pos = TRUE,logfc.threshold = 0.5)
GOmat <- clusterGO(markerlist)
GOheatmap(GOmat)
markerlist <- FindAllMarkers(pbmc,only.pos = TRUE,logfc.threshold = 0.5)
GOmat <- clusterGO(markerlist)
GOheatmap(GOmat)
}
