% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pathway_result.R
\name{GOplot}
\alias{GOplot}
\title{Dotplot of GO enrichment.}
\usage{
GOplot(
  GOmat,
  show.top.term = 5,
  log.adj.P = F,
  cell.order = NULL,
  term.order = NULL,
  coord_flip = F,
  color.use = NULL,
  text.y.size = 8,
  text.x.size = 10,
  legend.title.size = 10,
  legend.text.size = 10
)

GOplot(
  GOmat,
  show.top.term = 5,
  log.adj.P = F,
  cell.order = NULL,
  term.order = NULL,
  coord_flip = F,
  color.use = NULL,
  text.y.size = 8,
  text.x.size = 10,
  legend.title.size = 10,
  legend.text.size = 10
)
}
\arguments{
\item{GOmat}{Result of GO enrichment from clusterGO.}

\item{show.top.term}{Number of top GO terms of each cluster showed in dotplot.}

\item{log.adj.P}{Show -log(adjust p value) instead of adjust p value.}

\item{cell.order}{Cell order showed in dotplot.}

\item{term.order}{GO terms order showed in dotplot.}

\item{coord_flip}{Parameter in ggplot.}

\item{color.use}{Color patterns used for showing p value.}

\item{text.y.size}{Text size of GO terms.}

\item{text.x.size}{Text size of Celltypes.}

\item{legend.title.size}{Text size of legend title.}

\item{legend.text.size}{Text size of legend text.}
}
\description{
Dotplot of GO enrichment.

Dotplot of GO enrichment.
}
\examples{
markerlist <- FindAllMarkers(pbmc,only.pos = TRUE,logfc.threshold = 0.5)
GOmat <- clusterGO(markerlist)
GOplot(GOmat)
markerlist <- FindAllMarkers(pbmc,only.pos = TRUE,logfc.threshold = 0.5)
GOmat <- clusterGO(markerlist)
GOplot(GOmat)
}
