% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pathway_result.R
\name{GSEA_all_plot}
\alias{GSEA_all_plot}
\title{Bar plot to show the result of all clusters' GESA.}
\usage{
GSEA_all_plot(
  fgsea_df,
  only.pos = T,
  color.use = NULL,
  show.num = 15,
  pvalue = 0.05,
  text.gap = 0.1,
  title = NULL
)
}
\description{
Bar plot to show the result of all clusters' GESA.
}
\examples{
DEG_GSEA <- parallel::mclapply(levels(Idents(pbmc)), function(x) {
mmk <- FindMarkers(pbmc, ident.1 = x, test.use = 'wilcox',
                   min.pct = 0.25,logfc.threshold = 0)
mmk$gene <- rownames(mmk)
mmk$cluster <- x
mmk$cluster[mmk$avg_logFC < 0] <- paste0(x,'_Others')
mmk
}, mc.cores = 10)
names(DEG_GSEA) <- levels(Idents(pbmc))
gsealist <- GSEA_all(DEG_GSEA, category = 'C2',subcategory = NULL, n.cores = 10)
fgsea_df <- do.call(rbind, gsealist)
GSEA_all_plot(fgsea_df, show.num = 3)

}
