% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_enrichment.R
\name{GSEA_pair}
\alias{GSEA_pair}
\title{Pathway analysis of DEG from FindMarkers using fGSEA. Only consider TWO cell types.}
\usage{
GSEA_pair(
  DEGdf,
  species = "Homo sapiens",
  category = NULL,
  subcategory = NULL,
  simplify_pathways = T,
  return.enrichment.plot = F,
  enrichment.plot.pathways = NULL,
  eps = 0,
  nPermSimple = 1000
)
}
\arguments{
\item{DEGdf}{list of data frames from FindMarkers. DEGdf should include 3 columns: avg_logFC(the second celltype is negative), cluster(2 celltypes) and gene. Recommend logfc.threshold = 0.}

\item{species}{Species of input data.}

\item{category}{Category of MsigDB, such as C2. Use list_all_GSEA_categories() to see all category. See https://www.gsea-msigdb.org/gsea/msigdb/ for details.}

\item{subcategory}{Sub-class of the category, such as KEGG.}

\item{simplify_pathways}{Simplify the pathways names. Lower the alphabets and remove '_'.}

\item{eps}{Minimum of P-value. See ?fgseaMultilevel.}

\item{nPermSimple}{Number of permutations for estimation of P-values. See ?fgseaMultilevel.}
}
\description{
pathway – name of the pathway; pval – an enrichment p-value; padj – a BH-adjusted p-value; log2err – the expected error for the standard deviation of the P-value logarithm. ES – enrichment score, same as in Broad GSEA implementation; NES – enrichment score normalized to mean enrichment of random samples of the same size; size – size of the pathway after removing genes not present in 'names(stats)'. leadingEdge – vector with indexes of leading edge genes that drive the enrichment.
}
\examples{
celltype_1 <- 'CD8 T'
celltype_2 <- 'NK'
DEG_pair <- FindMarkers(pbmc, ident.1 = celltype_1, ident.2 = celltype_2, min.pct = 0.25, logfc.threshold = 0)
DEG_pair$gene <- rownames(DEG_pair)
DEG_pair$cluster <- celltype_1
DEG_pair$cluster[DEG_pair$avg_logFC < 0] <- celltype_2
GSEA_KEGG <- GSEA_pair(DEG_pair, category = 'C2', subcategory = 'KEGG')
GSEA_KEGG <- GSEA_KEGG[GSEA_KEGG$padj < 0.05, ]
GSEA_C2 <- GSEA_pair(DEG_pair, category = 'C2', subcategory = NULL)
GSEA_C2 <- GSEA_C2[GSEA_C2$padj < 0.05, ]

}
