% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pathway_result.R
\name{GSEA_pair_plot}
\alias{GSEA_pair_plot}
\title{pairs barplot showing the result of GESA.}
\usage{
GSEA_pair_plot(
  fgsea_df,
  show.num = 15,
  pvalue.adj = 0.05,
  text.gap = 0.1,
  text.max.length = 50,
  bar.width = 0.8,
  text.lineheight = 0.6,
  text.size = 3,
  show.gene.on.bar = F,
  gene.max.length = 10,
  gene.size = 3,
  gene.gap = 0.1,
  color = NULL,
  title = NULL,
  legend.title = NULL
)
}
\description{
pairs barplot showing the result of GESA.
}
\examples{
celltype_1 <- 'CD8 T'
celltype_2 <- 'NK'
DEG_pair <- FindMarkers(pbmc, ident.1 = celltype_1, ident.2 = celltype_2, min.pct = 0.25, logfc.threshold = 0)
DEG_pair$gene <- rownames(DEG_pair)
DEG_pair$cluster <- celltype_1
DEG_pair$cluster[DEG_pair$avg_logFC < 0] <- celltype_2

GSEA_KEGG <- GSEA_pair(DEG_pair, category = 'C2', subcategory = 'KEGG')
GSEA_KEGG <- GSEA_KEGG[GSEA_KEGG$padj < 0.05, ]
GSEA_C2 <- GSEA_pair(DEG_pair, category = 'C2', subcategory = NULL)
GSEA_C2 <- GSEA_C2[GSEA_C2$padj < 0.05, ]

GSEA_pair_plot(GSEA_C2)

}
