% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_highlight_cluster.R
\name{highallcl3}
\alias{highallcl3}
\title{highlight points of all cluster by separated UMAP in different colors.}
\usage{
highallcl3(
  SeuratS4,
  group.by = NULL,
  reductions = "umap",
  text.size = 8,
  pt.size = 0.1,
  AI.friendly = F,
  return.list = F,
  ncol = NULL,
  color = NULL
)
}
\arguments{
\item{SeuratS4}{A Seurat object.}

\item{reduction}{Celect the reduction method.}
}
\value{
A ggplot objects.
}
\description{
highlight points of all cluster by separated UMAP in different colors.
}
\examples{
data('pbmc')
highallcl(pbmc,reduction = "umap")
}
