% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_highlight_cluster.R
\name{highcl}
\alias{highcl}
\title{highlight points of one cluster}
\usage{
highcl(
  SeuratS4,
  cluster,
  reduction = "tsne",
  color.1 = "red",
  color.2 = "grey",
  top.layer = T,
  label = FALSE,
  label.size = 3,
  pt.size = 0.1,
  ...
)
}
\arguments{
\item{SeuratS4}{A Seurat object.}

\item{cluster}{Select which cluster to highlight.}

\item{reduction}{Select the reduction method.}

\item{top.layer}{If TRUE, put the cluster points on the top layer of plot.}

\item{label}{Parameter in Dimplot.}

\item{label.size}{Parameter in Dimplot.}

\item{pt.size}{Parameter in Dimplot.Set the point size.}

\item{...}{Parameter in Dimplot.}
}
\value{
A ggplot objects.
}
\description{
highlight points of one cluster
}
\examples{
data('pbmc')
highcl(pbmc,'Naive CD4 T',reduction = "umap")
}
