% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferCNV_preprocess.R
\name{inferCNV_preprocess}
\alias{inferCNV_preprocess}
\title{Function of pre-processing data for inferCNV}
\usage{
inferCNV_preprocess(
  object,
  meta.data,
  annotation,
  malignant_prefix = "^EP",
  group,
  min.cell.per.group = 30,
  gene.anno.gtf,
  outdir,
  subset.malig = F,
  subset.malig.num = 10000,
  subset.ref = T,
  subset.ref.prop = 0.5,
  seed = 123
)
}
\arguments{
\item{annotation}{Column of meta.data for distinguishing malignant and reference.}

\item{malignant_prefix}{Prefix of malignant names.}

\item{group}{Column of meta.data for annotation bar on the left side of CNV heatmap.}
}
\description{
Function of pre-processing data for inferCNV
}
\examples{
datalist <- inferCNV_preprocess(object, meta.data,annotation, malignant_prefix = '^EP',
group,min.cell.per.group = 30,gene.anno.gtf,outdir,subset.malig = F,
subset.ref = T,subset.ref.prop = 0.5,seed = 123)
infercnv_obj = CreateInfercnvObject(raw_counts_matrix=datalist$matrix,
annotations_file=datalist$annotations_file,delim="\t",
gene_order_file=datalist$gene_order_file,
ref_group_names = datalist$ref_group_names)
infercnv_obj = infercnv::run(infercnv_obj,cutoff=0.1, # cutoff=1 works well for Smart-seq2, and cutoff=0.1 works well for 10x Genomics
out_dir=outdir, cluster_by_groups=TRUE, denoise=TRUE,HMM=TRUE)
}
