% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dot.R
\name{inter_dotplot}
\alias{inter_dotplot}
\title{Merging two dotplots by gene-gene interaction lines.}
\usage{
inter_dotplot(
  nordata,
  ident,
  interdf,
  complex_input_path = NULL,
  add.complex.gene.names = F,
  highlight = c("no", "multicolor", "singlecolor"),
  multicolor.line = NULL,
  singlecolor.line.li = "red",
  singlecolor.line.re = "red",
  noSig.circle.color = "white",
  color.use = c("grey", "#111111", "black"),
  p_thresh = 0.01,
  ligandcell = "ligand",
  receptorcell = "receptor",
  ligandgene = "ligandgene",
  receptorgene = "receptorgene",
  cell.order.li = NULL,
  cell.order.re = NULL,
  gene.order.li = NULL,
  gene.order.re = NULL,
  circle.size.as = c("expression percent", "expressed cell number"),
  na.as = 0.05,
  scale.min = NA,
  scale.max = NA,
  do.scale.circle.size = T,
  coord_flip = F,
  line.size = 0.5,
  text.x.angle = 45,
  x.vjust = 0,
  x.hjust = 0,
  text.y.angle = 0,
  y.vjust = 0.5,
  y.hjust = 0.5,
  text.size = 8,
  text.y.size = 8,
  text.x.size = 10,
  legend.title.size = 10,
  x_position.li = "top",
  x_position.re = "top",
  legend.text.size = 8,
  legend.key.size = 1,
  legend.box.li = "horizontal",
  legend.box.re = "vertical",
  legend.title.li = "Ligand\\nExpression",
  legend.title.re = "Receptor\\nExpression",
  legend.position.li = "bottom",
  legend.position.re = "right",
  width = c(4, 1, 4),
  plot.margin.li = unit(c(0.1, 0.1, 0.1, 0.1), "cm"),
  plot.margin.re = unit(c(0.1, 0.1, 0.1, 0.1), "cm"),
  ...
)
}
\arguments{
\item{nordata}{A normalize matrix of gene expression.}

\item{ident}{A named vector/factor. The identity of cells. Could be Idents(object).}

\item{interdf}{A dataframe including 4 columns: ligandcell,receptorcell,ligandgene and receptorgene.}

\item{ligandcell}{Column name of interdf.}

\item{receptorcell}{Column name of interdf.}

\item{ligandgene}{Column name of interdf.}

\item{receptorgene}{Column name of interdf.}

\item{cell.order.li}{Cell order of ligand cells.}

\item{cell.order.re}{Cell order of receptor cells.}

\item{gene.order.li}{Gene order of ligand genes.}

\item{gene.order.re}{Gene order of receptor genes.}

\item{na.as}{Replace NA as numeric if NA exists.}

\item{width}{The relative widths to arrange 3 ggplot objects: heatmap + links + heatmap.}

\item{do.scale}{Scale gene expression to range of 0-1.}
}
\value{
ggplot object.
}
\description{
Merging two dotplots by gene-gene interaction lines.
}
\examples{
library(reshape2)
library(ggplot2)
library(dplyr)
ov_PT <- subset(biospiper::ov_obj, Group_abb == 'PT')
datalist_PT <- RunCellphoneDB(object = ov_PT,annotation = 'celltype',outdir = './test/PT')
#file_mean <- read.table(file='./test/PT/means.txt', header=T, sep='\t', check.names=F, as.is=T,stringsAsFactors = F)
#file_pval <- read.table(file='./test/PT/pvalues.txt', header=T, sep='\t', check.names=F, as.is=T,stringsAsFactors = F)
#datalist_PT <- list(mean = file_mean, pval = file_pval)
df_PT <- interaction_data(datalist_PT$mean, datalist_PT$pval,cellchat.evidence = T)
chemo_li <- c(rownames(ov_PT)[grepl('^CCL',rownames(ov_PT))], rownames(ov_PT)[grepl('^CXCL',rownames(ov_PT))],'XCL1','XCL2','CX3CL1')
df_PT <- df_PT[df_PT$ligandgene \%in\% chemo_li, ]
df_PT <- df_PT[, c('ligand','receptor','ligandgene','receptorgene','pvalue')]
heatmapcolor <- c('#4493c2','white','#f5b08f','#b93032','#b93032','#b93032','#b93032')
color.use <- grDevices::colorRampPalette(heatmapcolor)(10)
#No highlight
p <- inter_dotplot(ov_PT@assays$RNA@data, Seurat::Idents(ov_PT), df_PT,
highlight= 'no', width = c(6,1,5), legend.box.li = 'vertical', color.use = color.use)
#highlight by single color
p <- inter_dotplot(ov_PT@assays$RNA@data, Seurat::Idents(ov_PT), df_PT,
highlight= 'singlecolor', width = c(6,1,5), legend.box.li = 'vertical')
#highlight by multicolor
p <- inter_dotplot(ov_PT@assays$RNA@data, Seurat::Idents(ov_PT), df_PT,
highlight= 'multicolor', width = c(6,1,5), legend.box.li = 'vertical'))

}
