% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dot.R
\name{interaction_dotplot2}
\alias{interaction_dotplot2}
\title{dotplot}
\usage{
interaction_dotplot2(
  file_m,
  file_p,
  ligandcell = NULL,
  receptorcell = NULL,
  receptorcell.prefix = NULL,
  only.significant = T,
  min.p.value = 0.05,
  order.cells = NULL,
  sort.receptorcell = F,
  order.genepair.auto = F,
  object = NULL,
  annotation = NULL,
  order.genepair = NULL,
  max.cell = Inf,
  return.data = F,
  save.pdf = NULL,
  scale.pt.size = c(0, 3, 6),
  size.axis.text = 20,
  size.axis.text.x = 13,
  size.legend.title = 20,
  size.legend.text = 20,
  ...
)
}
\description{
dotplot
}
\examples{
datalist <- RunCellphoneDB(object = ov_obj,annotation = 'celltype',outdir = './test')
plist <- interaction_dotplot(datalist$mean, datalist$pval, ligandcell = 'DC',receptorcell.prefix = 'CD')
}
