% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PHASE1_run_Seurat_v5_QC_clustering.R
\name{is_outlier}
\alias{is_outlier}
\title{Define outlier by MAD. Refer to https://www.sc-best-practices.org/preprocessing_visualization/quality_control.html}
\usage{
is_outlier(metadata, metric, nmads = 5, mad.scale.factor = 1)
}
\description{
Define outlier by MAD. Refer to https://www.sc-best-practices.org/preprocessing_visualization/quality_control.html
}
