% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data_for_shiny.R
\name{iSEEplot}
\alias{iSEEplot}
\title{Visualize single cell RNA-Seq data in HTML format.}
\usage{
iSEEplot(data)
}
\arguments{
\item{data}{A SingleCellExperiment object.}
}
\description{
Visualize single cell RNA-Seq data in HTML format.
}
\examples{
data('pbmc')
pbmc@reductions$pca <- NULL
m_sce <- as.SingleCellExperiment(pbmc)
SingleCellExperiment::counts(m_sce) <- pbmc@assays$RNA@data
m_sce@colData$Annotation <- Idents(pbmc)
#iSEEplot(m_sce)
}
