% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_angrycell_result.R
\name{List_predicted_celltypes}
\alias{List_predicted_celltypes}
\title{Show predicted cell types in tree。}
\usage{
List_predicted_celltypes(
  angrycell.tab,
  outfile = NULL,
  type = "text",
  root = "Cell types"
)
}
\description{
Show predicted cell types in tree。
}
