% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_score.R
\name{M1M2_score}
\alias{M1M2_score}
\title{Define M1/M2 phenotypes of Myeloid cells as mean expression of gene signatures, and plot data.
M1/M2 markers were from DOI: 10.1016/j.cell.2020.03.048 and DOI: 10.1016/j.cell.2018.05.060}
\usage{
M1M2_score(
  object,
  Macrophage.names,
  return.data = F,
  M1M2list = NULL,
  AddModuleScore = F,
  plot.type = c("violin", "violin merge", "box", "correlation"),
  celltype.order = NULL,
  order.by.M1M2 = c("none", "M1", "M2", "diff"),
  correlation_centre = T,
  remove.axis.x = F,
  legend.title = "Average expression",
  legend.position = "right",
  legend.title.size = 8,
  strip.position = "bottom",
  nrow = 1,
  strip.text.size = 8,
  strip.text.angle = 45,
  strip.text.hjust = 1,
  strip.text.vjust = 1,
  panel.spacing = 0.5,
  clip = "on",
  border.color = "black",
  border.size = 0.5
)
}
\arguments{
\item{object}{A Seurat object}

\item{Macrophage.names}{A vector of all the Macrophage sub-type names.}

\item{return.data}{Return statistic data.}

\item{plot.type}{The type of plot. Only 'violin', 'violin merge','box' or 'correlation' is provided.}
}
\value{
A ggplot object of violin, box or heatmap.
}
\description{
Define M1/M2 phenotypes of Myeloid cells as mean expression of gene signatures, and plot data.
M1/M2 markers were from DOI: 10.1016/j.cell.2020.03.048 and DOI: 10.1016/j.cell.2018.05.060
}
