% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion_statistics.R
\name{MASC_multi}
\alias{MASC_multi}
\title{MASC - Mixed effect modeling of Associations of Single Cells}
\usage{
MASC_multi(
  dataset,
  cluster,
  contrast,
  model = c("2x2", "pairs"),
  random_effects = NULL,
  fixed_effects = NULL,
  adjust.method = "BH",
  verbose = FALSE,
  save_models = FALSE,
  save_model_dir = NULL,
  ncores = 4
)
}
\description{
MASC - Mixed effect modeling of Associations of Single Cells
}
\examples{
stat2x2 <- MASC_multi(meta200, cluster = 'maingroups', contrast = 'Group',
random_effects = 'Sample',model = '2x2')
stat_pairs <- MASC_multi(meta200, cluster = 'maingroups', contrast = 'Group',
random_effects = 'Sample',model = 'pairs')
}
