% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_merge.R
\name{mergeplot}
\alias{mergeplot}
\title{Gradual color plot of two co-expression genes.}
\usage{
mergeplot(xy, exprmat)
}
\arguments{
\item{xy}{cell location from seurat_object@reduction$tsne@cell.embeddings.}

\item{exprmat}{a normalize 2 selected genes expression dataframe(cell x gene).}
}
\description{
Gradual color plot of two co-expression genes.
}
\examples{
data('pbmc')
marker_to_plot <- c("CD4","CCR7")
exp <- log2(as.data.frame(t(pbmc@assays$RNA@data[marker_to_plot,]))+1)
xy <- pbmc@reductions$umap@cell.embeddings
mergeplot(xy,exp)
}
