% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_merge.R
\name{mergeplot1}
\alias{mergeplot1}
\title{3 color plot of two co-expression genes.}
\usage{
mergeplot1(SeuratS4, marker_to_plot, active.assay = "RNA", reduction = "umap")
}
\arguments{
\item{SeuratS4}{a Seurat object.}

\item{marker_to_plot}{two Genes to plot.}
}
\description{
3 color plot of two co-expression genes.
}
\examples{
data('pbmc')
marker_to_plot <- c("CD4","CCR7")
mergeplot1(pbmc,marker_to_plot)
}
