% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/db_manager_and_explore.R
\docType{data}
\name{meta200}
\alias{meta200}
\title{meta data}
\format{
An object of class \code{data.frame} with 200 rows and 11 columns.

An object of class \code{data.frame} with 200 rows and 11 columns.
}
\usage{
meta200

meta200
}
\description{
meta data

meta data
}
\examples{
meta200 <- data('meta200')
meta200 <- data('meta200')
}
\keyword{datasets}
