% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MSigDBlist}
\alias{MSigDBlist}
\title{To list all pathways in MsigDB v7.1.}
\format{
An object of class \code{list} of length 18.
}
\usage{
MSigDBlist
}
\value{
List of all pathways.
}
\description{
To list all pathways in MsigDB v7.1.
}
\examples{
MSigDBlist <- data('MSigDBlist')
}
\keyword{datasets}
