% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{order_heatmap_gene}
\alias{order_heatmap_gene}
\title{Order genes for plot.}
\usage{
order_heatmap_gene(
  object,
  features,
  assay = "RNA",
  slot = "data",
  min.pct = 0,
  method = "DEG",
  sort.name.per.group = F,
  name.decreasing = F,
  sort.name.per.group.byAnnot = F,
  gene.Annot = NULL,
  seriate.method = c("BEA_TSP", "BEA", "PCA", "PCA_angle"),
  ident = Idents(object),
  cell.order = NULL
)
}
\arguments{
\item{method}{DEG or seriation. DEG: Order genes by DEG; seriation: Order genes by R package seriation. Default is DEG.}

\item{gene.Annot}{Data.frame, included 2 columns: annotation and gene. If sort.name.per.group.byAnnot is true, order genes by gene.Annot data frame in each cluster.}

\item{seriate.method}{Methods only for seriation. Options: BEA_TSP, BEA, PCA, PCA_angle. Default is BEA_TSP.}
}
\description{
Order genes for plot.
}
\examples{
marker_to_plot <- c('CD3D','CD4','CD8A','CCR7','IL7R','NCAM1','CD1C','CD68','CD14','CD79A','PPBP','FCGR3A')
ordered_gene <- order_heatmap_gene(pbmc, marker_to_plot, method = 'DEG')
DotPlot(pbmc, features = ordered_gene)

}
