% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R
\name{pairwisede}
\alias{pairwisede}
\title{Find pair DEG.}
\usage{
pairwisede(
  seurat,
  ident.1 = NULL,
  ident.2 = NULL,
  group.by = NULL,
  slot.use = "data",
  assay = "RNA",
  test.use = "t"
)
}
\description{
Find pair DEG.
}
