% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/db_manager_and_explore.R
\docType{data}
\name{pbmc}
\alias{pbmc}
\title{seurat object}
\format{
An object of class \code{Seurat} with 13714 rows and 2638 columns.

An object of class \code{Seurat} with 13714 rows and 2638 columns.
}
\usage{
pbmc

pbmc
}
\description{
a seurat object for example

a seurat object for example
}
\examples{
pbmc <- data('pbmc')
pbmc <- data('pbmc')
}
\keyword{datasets}
