% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_process.R
\name{PCs_contribution}
\alias{PCs_contribution}
\title{Etimate the Score of PCs accounted for genes.}
\usage{
PCs_contribution(
  object,
  geneset = "all",
  species = "HomoSapiens",
  return.plot = T,
  return.data = F
)
}
\arguments{
\item{object}{A seurat object.}

\item{geneset}{Input gene set.Default is 'all',representing mt,ribosome,
and HSP.}

\item{species}{Species. Only Human and Mouse are permitted.}

\item{return.plot}{If ture, return ggplot object.}

\item{return.data}{If ture, return PCs x Score plotdata.}
}
\value{
ggplot object or plotdata.
}
\description{
Etimate the Score of PCs accounted for genes.
}
\examples{
PCs_contribution(pbmc)
PCs_contribution(pbmc,geneset = c('CD1C','CD14'))
}
