% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue_similar.R
\name{permutation_test_tissue}
\alias{permutation_test_tissue}
\title{Permutation test by shuffling tissue labels.}
\usage{
permutation_test_tissue(
  object,
  tissue_colnames,
  query_group,
  pcamat,
  k.out,
  k = 30,
  seed = 123,
  n.perm = 1000,
  pval.cutoff = 0.05,
  n.core = 4
)
}
\description{
Permutation test by shuffling tissue labels.
}
