% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PHASE1_run_Seurat_v3_QC_clustering.R
\name{PHASE1_run_Seurat_v3_QC_clustering}
\alias{PHASE1_run_Seurat_v3_QC_clustering}
\title{Run QC and clustering using Seurat v3.}
\usage{
PHASE1_run_Seurat_v3_QC_clustering(param)
}
\description{
Run QC and clustering using Seurat v3.
}
