% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_paired_pie.R
\name{pie_plot_group}
\alias{pie_plot_group}
\title{Function for paired pie plot}
\usage{
pie_plot_group(
  meta,
  cellID,
  sctypes,
  group.colname,
  group.names = c("intact gut", "leaky gut"),
  min.pct = 0.5,
  test.use = c("none", "t", "wilcox"),
  adjust.method = c("BH", "holm", "hochberg", "hommel", "bonferroni", "BY", "fdr",
    "none"),
  plabel = c("p.adj.signif", "p.adj", "p"),
  colors = NULL,
  legend.title = NULL,
  change.anno.df = NULL,
  cell.order = NULL
)
}
\arguments{
\item{meta}{Data frame including group and element scores.}

\item{cellID}{Select rows by rownames}

\item{sctypes}{colnames that are used for pie elements.}

\item{group.colname}{colname of group.}

\item{group.names}{2 group names included in group colname.}

\item{min.pct}{minmum percentage that determine to change names of elements to 'Others'.}

\item{test.use}{Whether to assign asterisks to legend text.}

\item{change.anno.df}{A data frame to adjust element names.}
}
\description{
Function for paired pie plot
}
