% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion_statistics.R
\name{plot_diver}
\alias{plot_diver}
\title{boxplot}
\usage{
plot_diver(
  df,
  pairwise_t,
  plabel = c("p.adj.signif", "p.adj"),
  add = NULL,
  xlab = NULL,
  ylab = NULL
)
}
\description{
boxplot
}
