% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar.R
\name{plot_fraction2}
\alias{plot_fraction2}
\title{Multi barplot to show fraction of cell number based on total number or user definition.}
\usage{
plot_fraction2(
  meta,
  x = colnames(meta)[1:(ncol(meta) - 1)],
  fill.bar = colnames(meta)[ncol(meta)],
  based.on = NULL,
  normailize.factor = 1,
  fraction.data = NULL,
  color.use = NULL,
  coord_flip = T,
  legend.pos = "top",
  x.size = 8,
  x.angle = 0,
  x.label = "",
  y.label = "",
  order.x.by.size = F,
  order.x.reverse = T,
  savedata = F,
  printdata = F,
  fill.bar.order = NULL
)
}
\arguments{
\item{meta}{A metadata contains data.fill(like celltypes),and data.x(like Groups, Samples and others metadata).}

\item{x}{Colnames of metadata as x axis. if x > 1, Default to use facet to merge boxplots.}

\item{fill.bar}{colname of metadata as fill data.}

\item{based.on}{colname of Alternative demominator instead of total count.}

\item{color.use}{fill color of bar. Default is 30 colors.}

\item{coord_flip}{exchange x,y axis.}

\item{legend.pos}{legend position. Available options: top, bottom, right and left.}

\item{x.size}{x text size.}

\item{x.angle}{x text angle.}

\item{x.label}{x axis name.}

\item{y.label}{y axis name.}

\item{savedata}{path to save data.}

\item{printdata}{print data.}
}
\value{
A ggplot objects (and data).
}
\description{
Multi barplot to show fraction of cell number based on total number or user definition.
}
\examples{
data('meta200')
meta <- meta200[,c('Group','Sample','Tissue','maingroups')]
plot_fraction(meta)
plot_fraction(meta, x = c('Group','Sample'), fill.bar = 'maingroups')
plot_fraction(meta, x = 'Group', fill.bar = 'maingroups', coord_flip = FALSE,
 legend.pos = 'right', x.angle = 90)
}
