% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Roe.R
\name{plot_Roe}
\alias{plot_Roe}
\title{Plot Ro/e heatmap}
\usage{
plot_Roe(
  meta,
  colname,
  rowname,
  cluster_col = F,
  cluster_row = T,
  max.Roe = 2.5,
  font_size = 12,
  x.size = 10,
  y.size = 10,
  geom_text_size = 11
)
}
\description{
Plot Ro/e heatmap
}
