% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_similarity.R
\name{Plot_similar}
\alias{Plot_similar}
\title{Plot result from Find_similar function.}
\usage{
Plot_similar(
  datalist,
  k_clusters = 4,
  layout = c("c.dend", "h.dend", "heatmap", "heatmap_cc"),
  clustering_distance = c("cormat", "pearson"),
  clustering.method = "complete",
  clustering_distance_cols = "correlation",
  clustering_distance_rows = "correlation",
  circle.text.size = 2.2,
  circle.line.size = 1,
  circle.margin = 3,
  label.color = NULL,
  per_label_color = NULL,
  dend.color = NULL,
  heatmap.display.numbers = T,
  ...
)
}
\description{
Plot result from Find_similar function.
}
\examples{
myeloid.integrated <- readRDS('/BGFS1/home/huangzh/workspace/project/ovarian_cancer/result/RNA/614.Zhang_Colon_10X_macrophage_12625cells/myeloid.integrated.rds')
datalist <- Find_similar(integrated.object = myeloid.integrated, assay = 'integrated',
var.genes.method = 'var', Annotation = 'Annotation',k_clusters = 2, plot = F, return.data = T)
#correlation heatmap
Plot_similar(datalist, layout = 'heatmap',heatmap.display.numbers = T)
Plot_similar(datalist, layout = 'heatmap',heatmap.display.numbers = F)
#dendrogram
Plot_similar(datalist, layout = 'h.dend')
# cirlize denrogram
Plot_similar(datalist, layout = 'c.dend',k_clusters = 6)

}
