% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_venn_TCR_share.R
\name{plot_stat}
\alias{plot_stat}
\title{Venn plot + bar plot of shared TCR}
\usage{
plot_stat(
  meta,
  group,
  clonotype,
  alpha = 0.5,
  margin = 0.2,
  main = NULL,
  vennfilename = NULL,
  ...
)
}
\description{
Venn plot + bar plot of shared TCR
}
