% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap_TCR_share.R
\name{plot_TCR_heatmap_archive}
\alias{plot_TCR_heatmap_archive}
\title{plot TCR heatmap from pre-processed data}
\usage{
plot_TCR_heatmap_archive(
  data,
  cell.order = NULL,
  widths = c(1, 8, 6),
  x.size = 15,
  legend.heatmap.size = 10,
  legend.heatmap.title.size = 15,
  add.legend.anno = T,
  only.legend.anno = F,
  legend.ncol = NULL,
  legend.anno.size = 10,
  legend.anno.title.size = 15
)
}
\description{
plot TCR heatmap from pre-processed data
}
