% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar.R
\name{plotbar}
\alias{plotbar}
\title{Barplot of fraction of cell number. Only sample and annotation were input, without groups.}
\usage{
plotbar(metaf, color = NULL, axis.line.size = 0.1)
}
\description{
Barplot of fraction of cell number. Only sample and annotation were input, without groups.
}
\examples{
metaf <- meta200[,c('Sample','Annotation')]
plotbar(metaf)

}
