% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_box.R
\name{plotbox}
\alias{plotbox}
\title{Boxplot of fraction of cell number.}
\usage{
plotbox(
  metaf,
  color = NULL,
  axis.line.size = 0.1,
  box.line.size = 0.1,
  outlier.size = 0.1
)
}
\description{
Boxplot of fraction of cell number.
}
\examples{
metaf <- meta200[,c('Sample','Group','Annotation')]
plotbox(metaf)

}
