% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_box.R
\name{plotbox2}
\alias{plotbox2}
\title{Boxplot of fraction of cell number vs total number of cells based on group.}
\usage{
plotbox2(
  metaf,
  total.number.based = "Sample",
  color = NULL,
  axis.line.size = 0.1,
  box.line.size = 0.1,
  outlier.size = 0.1
)
}
\description{
Boxplot of fraction of cell number vs total number of cells based on group.
}
