% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_box.R
\name{plotbox3}
\alias{plotbox3}
\title{Boxplot of fraction of cell number. Scale fraction in each cell type.}
\usage{
plotbox3(
  metaf,
  color = NULL,
  scale.frac = F,
  ylabel = "Fraction of cell number",
  axis.line.size = 0.1,
  box.line.size = 0.1,
  outlier.size = 0.1
)
}
\description{
Boxplot of fraction of cell number. Scale fraction in each cell type.
}
\examples{
metaf <- meta200[,c('Sample','Group','Annotation')]
plotbo3x(metaf)

}
