% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_box.R
\name{plotbox_facet}
\alias{plotbox_facet}
\title{Boxplot of fraction of cell number.}
\usage{
plotbox_facet(
  metaf,
  color = NULL,
  test = c("t", "wilcox"),
  alternative = c("two.sided", "great", "less"),
  adjust.method = c("BH", "holm", "hochberg", "hommel", "bonferroni", "BY", "fdr",
    "none"),
  labelp = c("p.adj", "p.adj.signif"),
  scale.frac = F,
  ref.group = NULL,
  ylim = NULL,
  label.expansion = c(0.05, 0.1),
  y.position = NULL,
  order.group = NULL,
  return.data = F,
  ...
)
}
\description{
Boxplot of fraction of cell number.
}
\examples{
metaf <- meta200[,c('Sample','Group','Annotation')]
plotbox_facet(metaf)

}
