% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_merge.R
\name{plotcogene}
\alias{plotcogene}
\title{Scater plot of two genes.}
\usage{
plotcogene(SeuratS4, cl, gene1, gene2, return.frac = F)
}
\arguments{
\item{SeuratS4}{a Seurat object.}

\item{cl}{select which cluster to plot.}

\item{gene1}{one of two genes.}

\item{gene2}{one of two genes.}
}
\description{
Scater plot of two genes.
}
\examples{
data('pbmc')
plotcogene(pbmc,'Naive CD4 T','CD4','CCR7')
}
