% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchR_plot_highlight.R
\name{plotEmbedding2}
\alias{plotEmbedding2}
\title{plot embedding}
\usage{
plotEmbedding2(
  ArchRProj = NULL,
  embedding = "UMAP",
  name = "Sample",
  AI.friendly = F,
  AI.script =
    "/BGFS1/projectdata/jasper/project/ovarian_cancer/shell/functions/ggAIplot.R",
  label.name = NULL,
  label.size = 4,
  color = NULL,
  cell.order = NULL,
  pt.size = 0.1,
  legend.position = "bottom",
  legend.text.size = 10,
  legend.point.size = 4,
  legend.nrow = 3
)
}
\description{
plot embedding
}
