% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchR_plot_gene.R
\name{plotgene_ArchR}
\alias{plotgene_ArchR}
\title{scatter plot of ArchR processed data}
\usage{
plotgene_ArchR(
  object,
  markerGenes,
  pt.size = 1,
  AI.friendly = F,
  AI.script =
    "/BGFS1/projectdata/jasper/project/ovarian_cancer/shell/functions/ggAIplot.R",
  ncol = ceiling(length(markerGenes)^0.5),
  plot.margin = unit(c(0, 0, 0, 0), "cm"),
  size.title = 10
)
}
\description{
scatter plot of ArchR processed data
}
