% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plotMASC}
\alias{plotMASC}
\title{Heatmap of result from MASC.}
\usage{
plotMASC(
  all_result,
  OR = "OR",
  display_numbers = T,
  label = "Pvalue",
  log2.transfer = F,
  legend.title = NULL,
  coord.flip = F
)
}
\description{
Heatmap of result from MASC.
}
\examples{
stat2x2 <- MASC_multi(meta200, cluster = 'maingroups', contrast = 'Group',
random_effects = 'Sample',model = '2x2')
plotMASC(stat2x2)

}
