% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plotOR}
\alias{plotOR}
\title{Heatmap to show result of fish exact test of cell proportion.}
\usage{
plotOR(
  testresult,
  OR = "OR",
  Status = "Group",
  Celltype = "Annotation",
  Pvalue = "adj.Pvalue",
  display_numbers = T,
  label = "Pstar",
  log2.transfer = F,
  negative.log10.transfer = F,
  rowname.order = NULL,
  colname.order = NULL,
  cut.off = NULL,
  breaks.hp = NA,
  legend.breaks = NA,
  legend.breaks.raw = F,
  legend.labels.raw = F,
  colors = NULL,
  color.ramp = NULL,
  mid.point.positive = 1,
  mid.point.negative = 0,
  legend.title = NULL,
  coord.flip = F,
  ...
)
}
\arguments{
\item{testresult}{A data frame from result of cell_test.}

\item{label}{What type of text filled in heatmap when display_numbers is TRUE.
\itemize{
 \item{Pstar}: display '*' to show significance;
 \item{OR}: display values from 'OR' column.
}}
}
\description{
Heatmap to show result of fish exact test of cell proportion.
}
\examples{
# 2x2
meta_epi <- meta200[meta200$maingroups == 'Epi_cells',]
statdf <- cell_test(meta_epi, Group = 'Group',Annotation = 'Annotation',
Sample = 'Sample', model = '2x2', test = 'fisher',alternative = 'two.sided',
adjust.method = 'BH')
# plot
plotOR(statdf, OR = 'OR', Status = 'Group', Celltype = 'Annotation', 
Pvalue = 'adj.Pvalue', display_numbers = F, label = 'Pstar',
log2.transfer = F, cut.off = c(-Inf,10), mid.point.positive = 1,coord.flip = F)
}
