% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Roe.R
\name{plotRO}
\alias{plotRO}
\title{Heatmap to show result of Ro/e matrix.}
\usage{
plotRO(
  rovemat,
  display_numbers = T,
  label = "Pstar",
  log2.transfer = F,
  negative.log10.transfer = F,
  paletteLength = 100,
  rowname.order = NULL,
  colname.order = NULL,
  cut.off = NULL,
  breaks.hp = NA,
  legend.breaks = NA,
  legend.breaks.raw = F,
  legend.labels.raw = F,
  colors = NULL,
  color.ramp = NULL,
  mid.point.positive = 1,
  mid.point.negative = 0,
  legend.title = NULL,
  coord.flip = F,
  ...
)
}
\arguments{
\item{label}{What type of text filled in heatmap when display_numbers is TRUE.
\itemize{
 \item{Pstar}: display '-/+' to show significance;
 \item{OR}: display values from 'OR' column.
}}
}
\description{
Heatmap to show result of Ro/e matrix.
}
\examples{
# plot
plotOR(statdf, OR = 'OR', Status = 'Group', Celltype = 'Annotation', 
Pvalue = 'adj.Pvalue', display_numbers = F, label = 'Pstar',
log2.transfer = F, cut.off = c(-Inf,10), mid.point.positive = 1,coord.flip = F)
}
