% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest.R
\name{predict_cell}
\alias{predict_cell}
\title{Predict cell types by random forest models or logistic regression}
\usage{
predict_cell(
  train_data,
  test_data,
  label.name,
  method = "clm",
  return.model = F
)
}
\arguments{
\item{train_data}{a cell x gene dataframe used for training, must have the same genes as test_data.}

\item{test_data}{a new cell x gene dataframe that user wants to predict cell types.}

\item{label.name}{the colname with labels in train_data}

\item{method}{Methods of prediction. Can be clm, rf_ranger, rf_rpart. Default is clm.}
}
\value{
A object result from stat::predict
}
\description{
Predict cell types by random forest models or logistic regression
}
