% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue_similar.R
\name{preprocess_netdata}
\alias{preprocess_netdata}
\title{Preprocess net data, return all connect cells.}
\usage{
preprocess_netdata(net_pos, meta, group, annotation, samples)
}
\description{
Preprocess net data, return all connect cells.
}
