% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue_similar.R
\name{preprocessing_obj}
\alias{preprocessing_obj}
\title{Preprocess Seurat object. Running PCA based on tissue specific HVGs.}
\usage{
preprocessing_obj(
  object,
  tissue_colnames,
  query_group,
  assay = "RNA",
  npcs = 50,
  hvg.method = "vst",
  hvg.nfeatures = 2000,
  vars.to.regress = NULL
)
}
\description{
Preprocess Seurat object. Running PCA based on tissue specific HVGs.
}
