% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_process.R
\name{QC_filtered}
\alias{QC_filtered}
\title{Statistic of the filtered matrix.}
\usage{
QC_filtered(SeuratS4, outdir, do_cellcycle = "TRUE")
}
\arguments{
\item{outdir}{The filedir to save statistic figures.}

\item{object}{A seurat object.}
}
\value{
A Seurat object.
}
\description{
Statistic of the filtered matrix.
}
\examples{
pbmc <- QC_filtered(pbmc,'./')
}
