% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_process.R
\name{QC_plot}
\alias{QC_plot}
\title{QC. Add percent.ribo or percent.HSP to object；return object or Ridgeplot.}
\usage{
QC_plot(
  SeuratS4,
  species = "Human",
  ident = NULL,
  add.line = F,
  minF = 200,
  maxF = 10000,
  percmt = 10,
  color.use = NULL,
  combine = T,
  return.obj = F
)
}
\description{
QC. Add percent.ribo or percent.HSP to object；return object or Ridgeplot.
}
\examples{
Idents(pbmc) <- pbmc$orig.ident
QC_plot(pbmc, species = 'Human', add.line = T,minF = 200, maxF = 10000, percmt = 10)
pbmc <- subset(pbmc, subset = nFeature_RNA > 800)
QC_plot(pbmc, species = 'Human', add.line = F)
Idents(pbmc) <- pbmc$Annotation
QC_plot(pbmc,species = 'Human', add.line = F)
##return object with percent.ribo or percent.HSP
pbmc <- QC_plot(pbmc, species = 'Human', return.obj = T)
}
