% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_process.R
\name{QC_raw}
\alias{QC_raw}
\title{Statistic of the orignal matrix.}
\usage{
QC_raw(
  object,
  outdir,
  species = "Human",
  do_cellcycle = "TRUE",
  plot.raw = "TRUE",
  max.mt = 10,
  min.features = 200,
  max.features = Inf,
  multi.samples = "TRUE"
)
}
\arguments{
\item{object}{A seurat object.}

\item{outdir}{The filedir to save statistic figures.}

\item{species}{Species of the sample. Only Human and Mouse are permitted.}

\item{multi.samples}{If TRUE, data contains samples from different individual.}
}
\value{
A Seurat object.
}
\description{
Statistic of the orignal matrix.
}
\examples{
pbmc <- QC_raw(pbmc,'./')
}
