% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_enrichment.R
\name{readGMT}
\alias{readGMT}
\title{Read gtm file from MSigDB and export data.frame for AngryCellDB.}
\usage{
readGMT(gmtfile)
}
\description{
Read gtm file from MSigDB and export data.frame for AngryCellDB.
}
