% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion_statistics.R
\name{rove}
\alias{rove}
\title{Statistic of Group-enrichment}
\usage{
rove(
  data = NULL,
  x,
  y,
  plot = FALSE,
  normalize = FALSE,
  display_numbers = F,
  ...
)
}
\arguments{
\item{data}{data contain x and y. Default = NULL}

\item{x}{Cell clusters.}

\item{y}{Cell groups.}
}
\description{
Statistic of Group-enrichment
}
\examples{
data('meta200')
rovemat <- rove(x=as.vector(meta200$Annotation),y=as.vector(meta200$Group),plot=TRUE)
}
