% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PHASE1_run_Seurat_v5_QC_clustering.R
\name{run_batch_effect_correction}
\alias{run_batch_effect_correction}
\title{Run Batch effect correction based on Incorporated methods in Seurat V5}
\usage{
run_batch_effect_correction(SeuratS4, param)
}
\description{
Run Batch effect correction based on Incorporated methods in Seurat V5
}
