% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_DEG.R
\name{run_DEG}
\alias{run_DEG}
\title{DEG methods for benchmarking single-cell RNA-seq data.}
\usage{
run_DEG(
  object,
  groupname,
  ident.1,
  ident.2 = NULL,
  covariates = NULL,
  min.pct = 0,
  logfc.threshold = 0,
  min.cell = 50,
  test.method = c("Seurat_wilcox", "Seurat_t", "Seurat_negbinom", "Seurat_poisson",
    "Seurat_DESeq2", "t_test", "poisson_Wald", "negbinom_Wald", "MAST", "scvi",
    "limma_trend", "limma_voom", "glmGamPoi", "DESeq2_LRT", "DESeq2_Wald", "edgeR_LRT",
    "edgeR_QLF", "Monocle3_quasipoisson", "Monocle3_negbinomial", "Monocle3_poisson",
    "Monocle3_gaussian", "Monocle3_zipoisson", "Monocle3_zinegbinomial",
    "Monocle3_mixed.negbinomial", "Monocle3_binomial", "SCDE", "miloDE", "DAseq_STG",
    "DAseq_wilcox", "DAseq_t", "DAseq_negbinom", 
     "DAseq_poisson", "DAseq_MAST",
    "DAseq_DESeq2", "Meld_rank", "Meld_t-test", "Meld_wald", "Meld_lrt",
    "Meld_Seurat_wilcox", "Meld_Seurat_t", "Meld_Seurat_negbinom", "Meld_Seurat_poisson",
    "Meld_Seurat_MAST", "Meld_Seurat_DESeq2"),
  pseudo_count = 1,
  adjust.method = c("bonferroni", "holm", "hochberg", "hommel", "BH", "BY", "fdr",
    "none"),
  n.cores = 1,
  workdir = NULL,
  python2use_Meld =
    "/media/london_B/huangzh/software/miniconda3/envs/FR-Perturb/bin/python",
  python2use_scvi =
    "/lustre/home/kwxiong/Huangzhihao/software/miniconda3/envs/scvi-env/bin/python",
  meld_script_path =
    "/lustre/home/kwxiong/Huangzhihao/project/benchmark_perturb/shell/Perturb_benchmark-main/script/Meld_all.py",
  meta_script_path =
    "/lustre/home/kwxiong/Huangzhihao/project/benchmark_perturb/shell/Perturb_benchmark-main/script/Meld_meta.py",
  deg_script_path =
    "/lustre/home/kwxiong/Huangzhihao/project/benchmark_perturb/shell/Perturb_benchmark-main/script/Meld_deg.py",
  classes_script_path =
    "/lustre/home/kwxiong/Huangzhihao/project/benchmark_perturb/shell/Perturb_benchmark-main/script/Meld_classes.py",
  scvi_script_path =
    "/lustre/home/kwxiong/Huangzhihao/project/benchmark_perturb/shell/Perturb_benchmark-main/script/scvi.py",
  ...
)
}
\arguments{
\item{object}{A Seurat object.}

\item{groupname}{Column name of meta.data including 2 groups to performing differential expression.}

\item{ident.1}{One cluster in groupname.}

\item{ident.2}{Another cluster in groupname.}

\item{covariates}{Column names in meta.data for confounding effect correction.}

\item{min.pct}{Minimum fraction of min.pct cells in either of the two populations in groupname.}

\item{logfc.threshold}{Minimum log2 Fold-change value}

\item{test.method}{test method to use.}

\item{pseudo_count}{pseudo count for log normalization in some method (DESeq2).}

\item{adjust.method}{Adjusted method for P-value.}

\item{n.cores}{Number of cores for parallel processes.}

\item{workdir}{Output directory for temporary result.}

\item{...}{Arguments passed to other methods and to specific DE methods.}

\item{pseudobulk}{Whether to perform pseudobulk anlaysis by Seurat::AggregateExpression.}

\item{pseudobulk_covar}{covariates for pseudobulk generation.}
}
\description{
DEG methods for benchmarking single-cell RNA-seq data.
}
\examples{
pbmc <- readRDS('../data/pbmc_v4.rds')
pbmc <- subset(pbmc, Annotation \%in\% c('B','NK'))
pbmc$Annotation <- factor(pbmc$Annotation, levels = unique(pbmc$Annotation))
# includes random effects
pbmc$batch <- c(rep('sample1',100), rep('sample2',100),rep('sample3',299))
resultlist <- run_DEG(pbmc, 'Annotation', 'B', 'NK', covariates = 'batch', test.method = c('Seurat_t','glmGamPoi'), min.pct = 0.3, logfc.threshold = 1)
resultlistp <- run_DEG(pbmc, 'Annotation', 'B', 'NK', pseudobulk = T, test.method = c('Seurat_wilcox','glmGamPoi'), pseudobulk_covar = 'batch',min.pct = 0.3, logfc.threshold = 1)
}
