% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_DEG_pseudo_libra.R
\name{run_DEG_pseudo_libra}
\alias{run_DEG_pseudo_libra}
\title{DEG analysis using libra.}
\usage{
run_DEG_pseudo_libra(
  object,
  groupname,
  ident.1,
  ident.2 = NULL,
  samplename = NULL,
  celltype = NULL,
  min.pct = 0,
  min.cells = 50,
  min_reps = 2,
  min_features = 0,
  de_family = "pseudobulk",
  de_method = c("DESeq2", "edgeR", "limma"),
  de_type = c("LRT", "QLF", "Wald", "trend", "voom"),
  adjust.method = c("fdr", "bonferroni", "holm", "hochberg", "hommel", "BH", "BY",
    "none"),
  n.cores = 1,
  ...
)
}
\arguments{
\item{object}{A Seurat object.}

\item{groupname}{Column name of meta.data including 2 groups to performing differential expression.}

\item{ident.1}{One cluster in groupname.}

\item{ident.2}{Another cluster in groupname.}

\item{min.pct}{Minimum fraction of min.pct cells in either of the two populations in groupname.}

\item{min_reps}{Argument to run_de. The minimum number of replicates in a cell type to retain it. Defaults to 2. description}

\item{de_family}{DEG family to use. Available options are: singlecell, pseudobulk and mixedmodel. Default is pseudobulk.}

\item{de_method}{DEG methods to use. Only 3 methods are available for psudobulk analysis: 'DESeq2','edgeR', 'limma'.}

\item{de_type}{Sub-types of DEG, including LRT, QLF, Wald, trend, voom.}

\item{adjust.method}{Adjusted method for P-value.}

\item{n.cores}{Number of cores for parallel processes.}

\item{...}{Arguments passed to other methods and to specific DE methods.}

\item{logfc.threshold}{Minimum log2 Fold-change value}

\item{test.method}{test method to use.}

\item{covariates}{latent variables for single-cell Seurat/Signac based methods.}
}
\description{
DEG analysis using libra.
}
\examples{
pbmc <- readRDS('../data/pbmc_v4.rds')
pbmc <- subset(pbmc, Annotation \%in\% c('B','NK'))
pbmc$Annotation <- factor(pbmc$Annotation, levels = unique(pbmc$Annotation))
# includes random effects
pbmc$batch <- c(rep('sample1',100), rep('sample2',100),rep('sample3',299))
DEG <- run_DEG_pseudo_libra(object = pbmc, celltype = 'orig.ident', samplename = 'batch', groupname = 'Annotation', ident.1 = 'B', ident.2 = 'NK')
}
