% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PHASE1_run_Seurat_v5_QC_clustering.R
\name{run_doublet_and_filter}
\alias{run_doublet_and_filter}
\title{Detect doublet and filter low-quality cells}
\usage{
run_doublet_and_filter(SeuratS4, param)
}
\description{
Detect doublet and filter low-quality cells
}
