% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_enrichKEGG_GSEA_GO.R
\name{run_GSEA}
\alias{run_GSEA}
\title{Gene set enrichment analysis}
\usage{
run_GSEA(
  markerlist,
  gene_colname = "gene",
  log2FC_colname = "avg_log2FC",
  filter.gene.pattern = NULL,
  geneType = "SYMBOL",
  OrgDb = "org.Hs.eg.db",
  eps = 1e-10,
  minSize = 15,
  maxSize = 500,
  return.gsea_obj = TRUE,
  return.symbol = TRUE,
  reduce.pathways = FALSE,
  padj.cutoff = 0.05,
  ...
)
}
\description{
Gene set enrichment analysis
}
