% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_monocle3.R
\name{run_monocle3}
\alias{run_monocle3}
\title{Run monocle 3}
\usage{
run_monocle3(
  object,
  annotation,
  initial.celltype = NULL,
  batch.effect = NULL,
  batch.effect.continuous = NULL,
  max_components = 2,
  use.seurat.preprocess = F,
  seurat.preprocess.method = NULL,
  use.seurat.reduction = F,
  seurat.embeddings.method = NULL,
  num_dim = 50,
  preprocess_method = "PCA",
  reduction_method = "UMAP",
  cluster_method = "leiden",
  k = 20,
  umap.metric = "cosine",
  umap.min_dist = 0.1,
  umap.n_neighbors = 15L,
  umap.fast_sgd = FALSE,
  umap.nn_method = "annoy",
  n.cores = 16,
  use_partition = T,
  close_loop = T,
  learn_graph_control = NULL,
  save.files = F,
  outdir = NULL,
  return.obj = T
)
}
\arguments{
\item{umap.min_dist}{Numeric indicating the minimum distance to be passed to
UMAP function. Default is 0.1.See uwot package's umap for details. 
umap.min_dist is smaller, points on UMAP are closer.}
}
\description{
Run monocle 3
}
\examples{
## Don't run
run_monocle3(object = pbmc, annotation = 'Annotation',
 initial.celltype = 'NK',
use.seurat.preprocess = F, use.seurat.reduction = T,
seurat.preprocess.method = NULL, seurat.embeddings.method = 'umap',
save.files = T, outdir = './', return.obj = F, n.cores = 16)
mono3 <- readRDS('./mono3.rds')
plot_cells(mono3,color_cells_by = 'Annotation',
label_groups_by_cluster=FALSE, label_leaves=FALSE, label_branch_points=FALSE)

}
