% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue_similar.R
\name{run_queryknn}
\alias{run_queryknn}
\title{Find knn among cells of all other tissues.}
\usage{
run_queryknn(pcamat, group_pos, other_pos, k = 30)
}
\description{
Find knn among cells of all other tissues.
}
