% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data_for_shiny.R
\name{sce_10000}
\alias{sce_10000}
\title{Shuffle cells for iSEE}
\usage{
sce_10000(
  SeuratS4 = SeuratS4,
  num = 1000,
  select.meta = c("nFeature_RNA", "nCount_RNA"),
  Assay_used = "RNA",
  reduction = "tsne"
)
}
\arguments{
\item{SeuratS4}{A Seurat object.}

\item{num}{Shuffle cell number.}

\item{select.meta}{metadata reserved.}
}
\value{
A SingleCellExperiment object.
}
\description{
Shuffle cells for iSEE
}
\examples{
data('pbmc')
pbmc@meta.data$Annotaion <- Idents(pbmc)
m_s4 <- sce_10000(SeuratS4 = pbmc, num = 1000, select.meta = c("Annotaion","nFeature_RNA",
"nCount_RNA"))
}
