% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_score.R
\name{score_correlation}
\alias{score_correlation}
\title{Function for M1M2_score.}
\usage{
score_correlation(
  object,
  genesetlist,
  cellpos,
  AddModuleScore = F,
  scatter_average = F
)
}
\description{
Scatter plot to show correlation of 2 geneset.
}
