% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{scPalette1}
\alias{scPalette1}
\title{Generate colors from a customed color palette. This function is from R package: CellChat.}
\usage{
scPalette1(n)
}
\arguments{
\item{n}{number of colors. Max number is 26.}
}
\value{
A color palette for plotting
}
\description{
Generate colors from a customed color palette. This function is from R package: CellChat.
}
