% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{scPalette2}
\alias{scPalette2}
\title{Generate colors from a customed color palette}
\usage{
scPalette2(n)
}
\arguments{
\item{n}{number of colors. Max number is 30.}
}
\value{
A color palette for plotting
}
\description{
Generate colors from a customed color palette
}
