% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{scPalette3}
\alias{scPalette3}
\title{Generate colors from https://doi.org/10.1016/j.cell.2020.03.048, Fig1A.}
\usage{
scPalette3(n)
}
\arguments{
\item{n}{number of colors. Max number is 44.}
}
\value{
A color palette for plotting
}
\description{
Generate colors from https://doi.org/10.1016/j.cell.2020.03.048, Fig1A.
}
