% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{scPalette4}
\alias{scPalette4}
\title{Generate ggplot2's default colors}
\usage{
scPalette4(n)
}
\arguments{
\item{n}{number of colors.}
}
\value{
A color palette for plotting
}
\description{
Generate ggplot2's default colors
}
