% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{scPalette_anno_light}
\alias{scPalette_anno_light}
\title{Generate colors of bar from https://doi.org/10.1016/j.cell.2020.03.048, Fig1A.}
\usage{
scPalette_anno_light(n)
}
\arguments{
\item{n}{number of colors. Max number is 44.}
}
\value{
A color palette for plotting
}
\description{
Generate colors of bar from https://doi.org/10.1016/j.cell.2020.03.048, Fig1A.
}
