% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{scPalette_heatmap_bar}
\alias{scPalette_heatmap_bar}
\title{Generate colors of heatmap bar.}
\usage{
scPalette_heatmap_bar(color = c("blue", "green")[1])
}
\arguments{
\item{color}{blue: red, white, blue; green: red, white, green.}
}
\value{
A color palette for plotting
}
\description{
Generate colors of heatmap bar.
}
