% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrublet_scanpy_R.R
\name{scrublet_scanpy_R}
\alias{scrublet_scanpy_R}
\title{Scrublet}
\usage{
scrublet_scanpy_R(
  seurat_obj,
  outdir,
  python_home = system("which python", intern = TRUE),
  source_py_script = NULL,
  expected_doublet_rate = 0.05,
  batch_key = "None",
  return_results_only = T
)
}
\arguments{
\item{seurat_obj}{the CellDataSet upon which to perform Scrublet}

\item{outdir}{Derectory to save the result and temp result of scrublet}

\item{python_home}{The python home directory where Scrublet is installed}

\item{source_py_script}{scrublet python script. default is script in inst/}

\item{expected_doublet_rate, }{float (optional, default=0.05), See scrublet reference - expected_doublet_rate: the 
fraction of transcriptomes that are doublets, typically 0.05-0.1. Results are not particularly sensitive to this parameter. For this example, the expected doublet rate comes from the Chromium User Guide: https://support.10xgenomics.com/permalink/3vzDu3zQjY0o2AqkkkI4CC}

\item{batch_key}{colname including sample names for batch correlation.}

\item{return_results_only}{bool (optional, default False)}
}
\value{
A data.frame including Cellname,doublet_score,predicted_doublet,Sample
}
\description{
See article doi: https://doi.org/10.1016/j.cels.2018.11.005
}
\examples{
pbmc$Sample <- 'single'
res <- scrublet_scanpy_R(pbmc, outdir = './test/')

}
