% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_process.R
\name{SE_evaluation}
\alias{SE_evaluation}
\title{Auto-select numbers of PCs and resolution in multi-sample data.}
\usage{
SE_evaluation(
  object,
  k.param = 30,
  nn.eps = 0,
  ndims = 100,
  algorithm = 1,
  select.PCs = NULL,
  normalize.entropy = F,
  plot.tree = F,
  return.list = F,
  select_resolution = NULL,
  reduction = "PC",
  ...
)
}
\arguments{
\item{object}{A seurat object.}

\item{k.param}{Defines k for the k-nearest neighbor algorithm.}

\item{ndims}{Total Number of PCs to compute and store (100 by default)}

\item{algorithm}{Parameter in FindClusters of Seurat.}

\item{select.PCs}{Input ranges of PCs for searching.}

\item{normalize.entropy}{Normalize Shannon Entropy of cell type fraction
in each cluster.}

\item{plot.tree}{If TURE, Creates a plot of a clustering tree showing
the relationship between clusterings at different resolutions.}

\item{return.list}{Return Shannon Entopy list instead of PCs and resolution.}

\item{select_resolution}{Input ranges of resolution for searching. for example: c(0.5, 2.5).}

\item{reduction}{PC, tsne or umap.}

\item{...}{Deliver parameters to get_numPCs_CUM.}

\item{plotfile}{the filepath of ElbowPlot file.}
}
\value{
A vector which contains predicted PCs and resolution.
}
\description{
Find cut off PC values by top 80% - 95% cumsum of PCstdev with gaps > 0.001,
and optimize the result by Shannon Entopy. The meta.data of object must contain "Sample".
}
\examples{
samples <- c(rep(c('sample1','sample2','sample3'),(ncol(pbmc) -1)/3),'sample1')
pbmc@meta.data$Sample <- samples
SE_evaluation(pbmc)
SE_evaluation(pbmc,normalize.entropy = T)
SE_evaluation(pbmc,select.PCs = c(8,15))
SE_evaluation(pbmc,select.PCs = c(8,15),normalize.entropy = T)
}
