% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap_TCR_share.R
\name{shareTCRdata_order}
\alias{shareTCRdata_order}
\title{order heatmap data and get share number of TCR for length of annotation bar.}
\usage{
shareTCRdata_order(mat)
}
\description{
order heatmap data and get share number of TCR for length of annotation bar.
}
