% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_angrycell_result.R
\name{show_result}
\alias{show_result}
\title{plot of gene and predicted cell types in GO terms}
\usage{
show_result(
  object = SeuratS4,
  active.assay = "RNA",
  reduction = "tsne",
  angrycell.tab = annotation,
  show.all = TRUE,
  return.grob = FALSE,
  seed = 123
)
}
\arguments{
\item{object}{A seurat object.}

\item{active.assay}{Default assay in seurat object which used to model.}

\item{reduction}{Select the reduction method.}

\item{angrycell.tab}{a dataframe of result of angrycell.}

\item{show.all}{Plot all predicted cell type and marker genes in GO terms.}

\item{return.grob}{return a list of grob results.}
}
\description{
plot of gene and predicted cell types in GO terms
}
\examples{
anno <- angrycell(pbmc, min.pct = 0.3, show.top = 10, use.definite.markers = F,
reduction = 'umap')
show_result(pbmc,reduction = 'umap',angrycell.tab = anno,
show.all = TRUE,return.grob = FALSE)
groblist <- show_result(pbmc,reduction = 'umap',angrycell.tab = anno,
show.all = FALSE,return.grob = TRUE)
#Because grid.draw only append a new layer on previous plot,
#use dev.off() before grid.draw.
grid::grid.draw(groblist[['B']])
}
