% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_process.R
\name{stat_num}
\alias{stat_num}
\title{Comparing number of cell before and after filtering.}
\usage{
stat_num(
  meta,
  min_nFeature_RNA = 200,
  max_nFeature_RNA = Inf,
  percent.mt = 10,
  min.param.list = NULL,
  max.param.list = NULL,
  str.param.list = NULL
)
}
\description{
Comparing number of cell before and after filtering.
}
\examples{
meta <- scPioneer::meta200[,c("Sample","percent.mt", "percent.ribo","nFeature_RNA","nCount_RNA")]
statmat <- stat_num(meta)
statmat <- stat_num(meta, min.param.list = list(nFeature_RNA = 200,nCount_RNA = 500),
max.param.list = list(percent.mt = 10, percent.ribo = 70))

}
