% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_process.R
\name{stat_table}
\alias{stat_table}
\title{Statistc of cell number before and after cell filtering.}
\usage{
stat_table(
  filtered_obj,
  raw_sample,
  percent.mt = "percent.mt",
  percent.ribo = "percent.ribo",
  nFeature_RNA = "nFeature_RNA",
  nCount_RNA = "nCount_RNA",
  Sample = "Sample"
)
}
\value{
A dataframe of cell number and other QC message.
}
\description{
Statistc of cell number before and after cell filtering.
}
\examples{
samples <- c(rep(c('sample1','sample2','sample3'),(ncol(pbmc) -1)/3),'sample1')
pbmc@meta.data$Sample <- samples
pbmc[["percent.ribo"]] <- PercentageFeatureSet(pbmc, pattern = "^RP[SL]")
raw_sample <- pbmc$Sample
pbmc_filtered <- subset(pbmc, subset = nFeature_RNA > 800)
QC_table <- stat_table(pbmc_filtered, raw_sample)

}
