% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_process.R
\name{stat_table_one_sample}
\alias{stat_table_one_sample}
\title{QC statistic for one sample}
\usage{
stat_table_one_sample(
  filtered_obj,
  raw_cell_num,
  percent.mt = "percent.mt",
  percent.ribo = "percent.ribo",
  nFeature_RNA = "nFeature_RNA",
  nCount_RNA = "nCount_RNA"
)
}
\description{
QC statistic for one sample
}
\examples{
samples <- c(rep(c('sample1','sample2','sample3'),(ncol(pbmc) -1)/3),'sample1')
pbmc@meta.data$Sample <- samples
pbmc[["percent.ribo"]] <- PercentageFeatureSet(pbmc, pattern = "^RP[SL]")
raw_sample <- pbmc$Sample
pbmc_filtered <- subset(pbmc, subset = nFeature_RNA > 800)
QC_table_one <- stat_table_one_sample(pbmc_filtered, ncol(pbmc))

}
