% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_process.R
\name{statnum}
\alias{statnum}
\title{statistic of cell number by table().}
\usage{
statnum(metadf, rowname, colname, col_annotation = NULL)
}
\value{
data frame
}
\description{
statistic of cell number by table().
}
\examples{
df <- statnum(meta200, rowname = 'Annotation', colname = 'Group')

}
