% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_process.R, R/subset.R
\name{subsetID}
\alias{subsetID}
\title{Extract cells by proportion}
\usage{
subsetID(IDmat, num = NULL, fraction = NULL, min.cell = 10, seed = 123)

subsetID(IDmat, num = NULL, fraction = NULL, min.cell = 10, seed = 123)
}
\description{
Extract cells by proportion

Extract cells by proportion
}
\examples{
pbmc$Cellname <- colnames(pbmc)
subsetID(pbmc@meta.data[,c('Annotation','Cellname')], num = 100)

pbmc$Cellname <- colnames(pbmc)
subsetID(pbmc@meta.data[,c('Annotation','Cellname')], num = 100)
}
