% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_process.R, R/subset.R
\name{subsetID2}
\alias{subsetID2}
\title{Extract cells by expected number}
\usage{
subsetID2(
  newI,
  expected.cell = 200,
  fraction = NULL,
  min.cell = 10,
  seed = 123
)

subsetID2(
  newI,
  expected.cell = 200,
  fraction = NULL,
  min.cell = 10,
  seed = 123
)
}
\description{
Extract cells by expected number

Extract cells by expected number
}
\examples{
subsetID2(newI, expected.cell = 200, seed = 123)

subsetID2(newI, expected.cell = 200, seed = 123)

}
