% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_venn_TCR_share.R
\name{t_col}
\alias{t_col}
\title{Transparent colors
Inspired by Mark Gardener 2015, www.dataanalytics.org.uk}
\usage{
t_col(color, percent = 50, name = NULL)
}
\description{
Transparent colors
Inspired by Mark Gardener 2015, www.dataanalytics.org.uk
}
