% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_manager_and_explore.R
\name{Template}
\alias{Template}
\title{Template of uploading markers for AngryCellDB v1.2.0}
\usage{
Template(version = "v1.2.0")
}
\description{
Template of uploading markers for AngryCellDB v1.2.0
}
\examples{
Template()
}
