% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pathway_result.R
\name{text_ellipsis}
\alias{text_ellipsis}
\title{Convert string to ellipsis.}
\usage{
text_ellipsis(string, gene.max.length = 5, sep = ",", replaceby = "...")
}
\description{
Convert string to ellipsis.
}
