% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_manager_and_explore.R
\name{upload_marker}
\alias{upload_marker}
\title{Upload markers for AngryCellDB v1.2.0}
\usage{
upload_marker(
  txtfilename,
  author = "anonymous",
  outdir = "/BGFS1/projectdata/jasper/database/marker_collection",
  version = NULL
)
}
\description{
Upload markers for AngryCellDB v1.2.0
}
\examples{
upload_marker('/BGFS1/projectdata/jasper/database/marker_collection/marker_collection_template_v1.2.0.txt', author = "ShaoYu Song")
}
