% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_manager_and_explore.R
\name{uploaddb}
\alias{uploaddb}
\title{upload marker}
\usage{
uploaddb(cellname, markers = "_", Definite_markers = "_", author = "User")
}
\arguments{
\item{cellname}{cell type name}

\item{markers}{general expressing genes.}

\item{Definite_markers}{specific markers of this cell type.}

\item{author}{user who uploads this data.}
}
\description{
upload marker
}
\examples{
uploaddb('CD4+ naive T cell', markers = c('CD4','LEF1','TCF7','CCR7','SELL'),
Definite_markers = c('CD4','CCR7'), author = 'Zhihao Huang')

}
