% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion_statistics.R
\name{vegan_diver}
\alias{vegan_diver}
\title{Shannon, Simpson, and Fisher diversity indices and species richness.}
\usage{
vegan_diver(
  meta,
  annotation = "Annotation",
  sample = "Sample",
  group = "Group",
  index = "simpson",
  melt.data = T
)
}
\description{
Shannon, Simpson, and Fisher diversity indices and species richness.
}
