% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_violin.R
\name{violin_facet_one_gene}
\alias{violin_facet_one_gene}
\title{Violin plot of one gene in different groups for each cell type.}
\usage{
violin_facet_one_gene(
  object,
  celltype,
  group,
  gene,
  assay = "RNA",
  facet.toward = "row",
  strip.text.size = 8,
  strip.text.angle = NULL,
  strip.text.hjust = NULL,
  strip.text.vjust = NULL,
  panel.spacing = 0.5,
  clip = "on",
  border.color = "black",
  border.size = 0.5,
  color.use = NULL,
  jitter = F,
  celltype.order = NULL,
  group.order = NULL,
  add.ave.point = T,
  ave.point.size = 1.5,
  add.line = T,
  line.size = 0.5,
  add.box = F,
  strip.text.switch = NULL,
  fill.expression = F,
  fill.color = brewer.pal(n = 9, name = "YlOrRd"),
  mid.color.point = "auto",
  legend.position = "top",
  legend.name = "",
  jitter.size = 0.1,
  text.x.size = 8,
  text.y.size = 8,
  hjust = 0,
  vjust = 0,
  legend.scale = NULL,
  legend.title.size = 8,
  axis.text.x.angle = NULL,
  remove.axis.x = F,
  remove.axis.y = F
)
}
\arguments{
\item{celltype}{A vector contains cell types (same as Idents(object));}

\item{group}{A vector contains group names (group of cell types).}

\item{gene}{Character. One gene name.}
}
\description{
Violin plot of one gene in different groups for each cell type.
}
