% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_score.R
\name{violin_score}
\alias{violin_score}
\title{Violin plot of gene score.}
\usage{
violin_score(
  object,
  geneset,
  genesetname,
  cellpos,
  AddModuleScore = F,
  border.color = "black",
  border.size = 1
)
}
\description{
Violin plot of gene score.
}
\examples{
cellpos <- pbmc$Annotation \%in\% c('Naive CD4 T', 'Memory CD4 T','CD8 T','NK')
geneset <- c('CCR7','TCF7','SELL')
violin_score(pbmc, geneset,'Naive lymphocytes',cellpos)

}
