% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_score.R
\name{violin_score2}
\alias{violin_score2}
\title{Violin plot of gene score.}
\usage{
violin_score2(
  object,
  geneset,
  genesetname,
  group.by = NULL,
  cell.order = NULL,
  assay = "RNA",
  slot = "data",
  scale = F,
  Add.boxplot = T,
  AddModuleScore = F,
  test.use = c("none", "t", "wilcox"),
  adjust.p = T,
  plabel = c("p.adj.signif", "p.adj", "p"),
  adjust.method = c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr",
    "none"),
  ref.group = NULL,
  remove.bracket = F,
  text.pval.size = 3.88,
  y.position.adjust = 0,
  color.use = NULL,
  midpoint = NULL,
  outlier.size = 0.5,
  violin.border.color = "black",
  box.border.color = "black",
  box.border.size = 0.5,
  return.stat = F,
  legend.name = NULL,
  ...
)
}
\description{
Violin plot of gene score.
}
\examples{
cellpos <- pbmc$Annotation \%in\% c('Naive CD4 T', 'Memory CD4 T','CD8 T','NK')
geneset <- c('CCR7','TCF7','SELL')
violin_score(pbmc, geneset,'Naive lymphocytes',cellpos)

}
