% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_score.R
\name{violin_score4}
\alias{violin_score4}
\title{Violin plot of gene score. Genes were averaged in each sample.}
\usage{
violin_score4(
  object,
  geneset,
  genesetname,
  group.by = NULL,
  facet.by = NULL,
  Add.boxplot = T,
  Add.jitter = F,
  jitter.size = 1,
  AddModuleScore = F,
  test.use = c("none", "t", "wilcox"),
  adjust.p = T,
  plabel = c("p.adj.signif", "p.adj", "p"),
  adjust.method = c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr",
    "none"),
  ref.group = NULL,
  violin.trim = T,
  remove.bracket = F,
  text.pval.size = 3.88,
  y.position.adjust = 0,
  adjust_p_text_gap = 1,
  adjust_p_text_height = 1,
  ylim = NULL,
  ylim.adjust = 1.1,
  color.use = NULL,
  midpoint = NULL,
  outlier.size = 0.5,
  legend.name = NULL,
  return.stat = F,
  ...
)
}
\description{
Violin plot of gene score. Genes were averaged in each sample.
}
