% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_violin.R
\name{vioplot2}
\alias{vioplot2}
\title{violin plot of gene expression.}
\usage{
vioplot2(
  object,
  paths,
  ident = Idents(object),
  assay = "RNA",
  select.celltypes = NULL,
  facet.toward = "row",
  do.log = F,
  strip.text.size = 8,
  strip.text.angle = NULL,
  strip.text.hjust = NULL,
  strip.text.vjust = NULL,
  strip.text.switch = NULL,
  border.color = "black",
  border.size = 0.5,
  color.use = NULL,
  jitter = F,
  cell.order = NULL,
  add.ave.point = F,
  add.line = F,
  line.size = 0.5,
  add.box = F,
  fill.expression = F,
  fill.color = c("#FEE090", "#D73027", "#A50026"),
  mid.color.point = "auto",
  legend.position = "top",
  legend.name = "",
  jitter.size = 0.1,
  text.x.size = 8,
  text.y.size = 8,
  hjust = 1,
  vjust = 1,
  legend.scale = NULL,
  legend.title.size = 8,
  axis.text.x.angle = 45,
  panel.spacing = 0.5,
  clip = "on",
  remove.axis.x = F,
  remove.axis.y = F,
  ...
)
}
\arguments{
\item{object}{A Seurat object.}

\item{paths}{Genes to plot.}

\item{ident}{Cell types or other meta data.}

\item{...}{Arguments passed to vioplot.}
}
\value{
A ggplot object.
}
\description{
See ?vioplot for R documentation.
}
\examples{
marker_to_plot <- c("MS4A1", "GNLY", "CD3E", "CD14", "FCER1A", "FCGR3A", "LYZ", "PPBP", "CD8A")
vioplot2(pbmc, marker_to_plot)
vioplot2(pbmc, marker_to_plot, fill.expression = F,add.ave.point = F,add.line = F,jitter = F,add.box = F)
vioplot2(pbmc, marker_to_plot, fill.expression = F,add.ave.point = F,add.line = F,jitter = F,add.box = T)
vioplot2(pbmc, marker_to_plot, fill.expression = F,add.ave.point = T,add.line = T,jitter = F,add.box = T)
vioplot2(pbmc, marker_to_plot, fill.expression = F,add.ave.point = T,add.line = T,jitter = T,add.box = T)
vioplot2(pbmc, marker_to_plot, fill.expression = T,add.ave.point = T,add.line = T,jitter = F,add.box = F)
vioplot2(pbmc, marker_to_plot, fill.expression = F,add.ave.point = T,add.line = T,jitter = T,add.box = T,
facet.toward = 'col')

}
