% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_violin.R
\name{vioplot3}
\alias{vioplot3}
\title{Violin plot of average expression of gene set.}
\usage{
vioplot3(object, pathlist, ident = Idents(object), assay = "RNA", ...)
}
\arguments{
\item{pathlist}{A list includes several gene sets.}
}
\value{
A ggplot object.
}
\description{
Violin plot of average expression of gene set.
}
