% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchR_plot_gene.R
\name{vioplot_ArchR}
\alias{vioplot_ArchR}
\title{violin plot of ArchR processed data}
\usage{
vioplot_ArchR(
  object,
  features,
  GroupBy = "Clusters",
  colorBy = "GeneScoreMatrix",
  ncol = 1,
  addBoxPlot = T,
  plot.data = T,
  return.data = F,
  ...
)
}
\description{
violin plot of ArchR processed data
}
