% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_violin.R
\name{vioplot_one_celltype}
\alias{vioplot_one_celltype}
\title{violin plot for one celltype}
\usage{
vioplot_one_celltype(
  object,
  paths,
  ident = Idents(object),
  assay = "RNA",
  facet.toward = "row",
  do.log = F,
  strip.text.size = 8,
  strip.text.angle = NULL,
  strip.text.hjust = NULL,
  strip.text.vjust = NULL,
  strip.text.switch = NULL,
  border.color = "black",
  border.size = 0.5,
  color.use = NULL,
  jitter = F,
  cell.order = NULL,
  add.ave.point = T,
  add.line = T,
  line.size = 0.5,
  add.box = F,
  fill.expression = F,
  fill.color = c("#FEE090", "#D73027", "#A50026"),
  mid.color.point = "auto",
  legend.position = "top",
  legend.name = "",
  jitter.size = 0.1,
  text.x.size = 8,
  text.y.size = 8,
  hjust = 0,
  vjust = 0,
  legend.scale = NULL,
  legend.title.size = 8,
  axis.text.x.angle = NULL,
  panel.spacing = 0.5,
  clip = "on",
  remove.axis.x = F,
  remove.axis.y = F,
  ave.point.size = 1.5,
  legend.key.size = 1,
  legend.text.size = 8
)
}
\description{
violin plot for one celltype
}
