% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{volcanoplot}
\alias{volcanoplot}
\title{Volcano plot}
\usage{
volcanoplot(
  DEmat,
  avg_logFC = "avg_logFC",
  p_val_adj = "p_val_adj",
  feature = "gene",
  color.use = c("red", "grey"),
  p_threshold = 0.05,
  FC_threshold = 0.25,
  topn = 5,
  show_gene = NULL,
  only.show.sign.text = T,
  text.size = 5,
  pt.size = 1,
  title = ""
)
}
\description{
Volcano plot
}
