% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{volcanoplot2}
\alias{volcanoplot2}
\title{Diagonal volcano plot}
\usage{
volcanoplot2(
  object,
  DEmat,
  annotation,
  ident.1,
  ident.2,
  assay = "RNA",
  avg_logFC_name = "avg_logFC",
  p_val_adj_name = "p_val_adj",
  legend_logFC_name = "logFC",
  legend_p_value_name = "FDR",
  feature_name = "gene",
  do.expm1 = F,
  color.use = c("red", "grey"),
  p_threshold = 0.05,
  FC_threshold = 0.25,
  show.gene = F,
  topn_p = NULL,
  topn_FC = NULL,
  add.gene.list = NULL,
  AI.friendly = F,
  text.size = 3,
  pt.size = 1,
  legend.point.size = 3,
  repel.max.overlaps = Inf,
  min.segment.length = 0.5,
  max.iter = 10000,
  title = "",
  ...
)
}
\description{
Diagonal volcano plot
}
