% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R, R/DEG_limma_trend.R
\name{DEG_limma_trend}
\alias{DEG_limma_trend}
\title{Pairwise DEG using limma-trend
# citation: https://ucdavis-bioinformatics-training.github.io/2019-single-cell-RNA-sequencing-Workshop-UCD_UCSF/scrnaseq_analysis/scRNA_Workshop-PART6.html}
\usage{
DEG_limma_trend(
  object,
  group.colname,
  ident.1,
  ident.2,
  batch.colname = NULL,
  min.pct = 0.2,
  logfc.threshold = 0,
  calcNormFactors = T,
  prior.count = 1,
  duplicateCorrelation.name = NULL,
  trend = T,
  adjust.method = c("bonferroni", "holm", "hochberg", "hommel", "BH", "BY", "fdr",
    "none"),
  sort.by = "P",
  ...
)

DEG_limma_trend(
  object,
  group.colname,
  ident.1,
  ident.2,
  batch.colname = NULL,
  min.pct = 0.2,
  logfc.threshold = 0,
  calcNormFactors = T,
  prior.count = 1,
  duplicateCorrelation.name = NULL,
  trend = T,
  adjust.method = c("bonferroni", "holm", "hochberg", "hommel", "BH", "BY", "fdr",
    "none"),
  sort.by = "P",
  ...
)
}
\arguments{
\item{group.colname}{Colname including two groups for DEG analysis, e.g. SNP and WT labels.}

\item{batch.colname}{Colname including covariate in meta.data.}

\item{calcNormFactors}{Calculate normalization factors from counts data.}

\item{...}{Parameter to GetAssayData.}

\item{min.cell}{Genes with number of expressed cells smaller than min.cell will be removed.}

\item{max.cpm.cut.off}{Filter low-expressed genes.}
}
\description{
Pairwise DEG using limma-trend
# citation: https://ucdavis-bioinformatics-training.github.io/2019-single-cell-RNA-sequencing-Workshop-UCD_UCSF/scrnaseq_analysis/scRNA_Workshop-PART6.html

Pairwise DEG using limma-trend
# citation: https://ucdavis-bioinformatics-training.github.io/2019-single-cell-RNA-sequencing-Workshop-UCD_UCSF/scrnaseq_analysis/scRNA_Workshop-PART6.html
# Refer to https://github.com/eonurk/RNA-seq-differential-analyses-guideline
}
\examples{
DEG <- DEG_limma_trend(pbmc, 'Annotation', ident.1 = 'Memory CD4 T', ident.2 = 'Naive CD4 T')
# includes random effects
pbmc$batch <- c(rep('sample1',1000), rep('sample2',1000),rep('sample3',638))
DEG <- DEG_limma_trend(pbmc, 'Annotation', ident.1 = 'Memory CD4 T', ident.2 = 'Naive CD4 T', batch.colname = 'batch')
pbmc <- readRDS('../data/pbmc_v4.rds')
DEG <- DEG_limma_trend(pbmc, 'Annotation', ident.1 = 'Memory CD4 T', ident.2 = 'Naive CD4 T')
# includes random effects
pbmc$batch <- c(rep('sample1',1000), rep('sample2',1000),rep('sample3',638))
DEG <- DEG_limma_trend(pbmc, 'Annotation', ident.1 = 'Memory CD4 T', ident.2 = 'Naive CD4 T', batch.colname = c('nCount_RNA','batch'))
}
