% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG.R, R/DEG_limma_voom.R
\name{DEG_limma_voom}
\alias{DEG_limma_voom}
\title{Pairwise DEG using limma-voom}
\usage{
DEG_limma_voom(
  object,
  group.colname,
  ident.1,
  ident.2,
  min.pct = 0.2,
  logfc.threshold = 0,
  batch.colname = NULL,
  adjust.method = c("bonferroni", "holm", "hochberg", "hommel", "BH", "BY", "fdr",
    "none"),
  duplicateCorrelation.name = NULL,
  plot.voom = F,
  ...
)

DEG_limma_voom(
  object,
  group.colname,
  ident.1,
  ident.2,
  min.pct = 0.2,
  logfc.threshold = 0,
  batch.colname = NULL,
  adjust.method = c("bonferroni", "holm", "hochberg", "hommel", "BH", "BY", "fdr",
    "none"),
  duplicateCorrelation.name = NULL,
  plot.voom = F,
  ...
)
}
\arguments{
\item{group.colname}{Colname including two groups for DEG analysis, e.g. SNP and WT labels.}

\item{batch.colname}{Colname including covariate in meta.data.}

\item{...}{Parameter to GetAssayData.}

\item{min.cell}{Genes with number of expressed cells smaller than min.cell will be removed.}

\item{max.cpm.cut.off}{Filter low-expressed genes.}
}
\description{
Pairwise DEG using limma-voom

Pairwise DEG using limma-voom.
Refer to https://ucdavis-bioinformatics-training.github.io/2018-June-RNA-Seq-Workshop/thursday/DE.html
}
\examples{
DEG <- DEG_limma_voom(pbmc, 'Annotation', ident.1 = 'Memory CD4 T', ident.2 = 'Naive CD4 T')
# includes random effects
pbmc$batch <- c(rep('sample1',1000), rep('sample2',1000),rep('sample3',638))
DEG <- DEG_limma_voom(pbmc, 'Annotation', ident.1 = 'Memory CD4 T', ident.2 = 'Naive CD4 T', batch.colname = 'batch')
pbmc <- readRDS('../data/pbmc_v4.rds')
DEG <- DEG_limma_voom(pbmc, 'Annotation', ident.1 = 'Memory CD4 T', ident.2 = 'Naive CD4 T')
# includes random effects
pbmc$batch <- c(rep('sample1',1000), rep('sample2',1000),rep('sample3',638))
DEG <- DEG_limma_voom(pbmc, 'Annotation', ident.1 = 'Memory CD4 T', ident.2 = 'Naive CD4 T', batch.colname = 'batch')
}
