% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDVA_DoHeatmapPlot.R
\name{DoHeatmapPlot}
\alias{DoHeatmapPlot}
\title{Do the geneset heatmap plot.}
\usage{
DoHeatmapPlot(
  plot.data,
  genes,
  group.by,
  color.panel,
  only.heatmap = F,
  coord.flip = F,
  return.obj = F,
  label.gene = NULL,
  break.gene = NULL,
  font.size.x = 15,
  font.size.y = 12,
  font.size.legend = 12,
  legend.name = ""
)
}
\arguments{
\item{plot.data}{The plot data created from getPlotData function.}

\item{genes}{The genes used in the plot.}

\item{group.by}{The cells should be ordered by or grouped by.}

\item{color.panel}{The color palette used when color by genes expression.}

\item{font.size}{The size of text, default by 15.}
}
\value{
A heatmap plot.
}
\description{
Do the geneset heatmap plot.
}
